Network.AWS.EMR

Description

Amazon Elastic MapReduce (Amazon EMR) is a web service that makes it easy to process large amounts of data efficiently. Amazon EMR uses Hadoop processing combined with several AWS products to do such tasks as web indexing, data mining, log file analysis, machine learning, scientific simulation, and data warehousing.

Documentation

module Network.AWS.EMR.AddInstanceGroups

module Network.AWS.EMR.AddJobFlowSteps

module Network.AWS.EMR.AddTags

module Network.AWS.EMR.DescribeCluster

module Network.AWS.EMR.DescribeJobFlows

module Network.AWS.EMR.DescribeStep

module Network.AWS.EMR.ListBootstrapActions

module Network.AWS.EMR.ListClusters

module Network.AWS.EMR.ListInstanceGroups

module Network.AWS.EMR.ListInstances

module Network.AWS.EMR.ListSteps

module Network.AWS.EMR.ModifyInstanceGroups

module Network.AWS.EMR.RemoveTags

module Network.AWS.EMR.RunJobFlow

module Network.AWS.EMR.SetTerminationProtection

module Network.AWS.EMR.SetVisibleToAllUsers

module Network.AWS.EMR.TerminateJobFlows

module Network.AWS.EMR.Types

module Network.AWS.EMR.Waiters