Network.AWS.EMR.AddInstanceGroups

Description

AddInstanceGroups adds an instance group to a running cluster.

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_AddInstanceGroups.html

Request

data AddInstanceGroups

Request constructor

addInstanceGroups

Arguments

:: Text
aigJobFlowId

-> AddInstanceGroups

AddInstanceGroups constructor.

The fields accessible through corresponding lenses are:

aigInstanceGroups :: [InstanceGroupConfig]

aigJobFlowId :: Text

Request lenses

aigInstanceGroups :: Lens' AddInstanceGroups [InstanceGroupConfig]

Instance Groups to add.

aigJobFlowId :: Lens' AddInstanceGroups Text

Job flow in which to add the instance groups.

Response

data AddInstanceGroupsResponse

Response constructor

addInstanceGroupsResponse :: AddInstanceGroupsResponse

AddInstanceGroupsResponse constructor.

The fields accessible through corresponding lenses are:

aigrInstanceGroupIds :: [Text]

aigrJobFlowId :: Maybe Text

Response lenses

aigrInstanceGroupIds :: Lens' AddInstanceGroupsResponse [Text]

Instance group IDs of the newly created instance groups.

aigrJobFlowId :: Lens' AddInstanceGroupsResponse (Maybe Text)

The job flow ID in which the instance groups are added.