Network.AWS.EMR.AddTags

Description

Adds tags to an Amazon EMR resource. Tags make it easier to associate clusters in various ways, such as grouping clusters to track your Amazon EMR resource allocation costs. For more information, see Tagging Amazon EMRResources.

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_AddTags.html

Request

data AddTags

Request constructor

addTags

Arguments

:: Text
atResourceId

-> AddTags

AddTags constructor.

The fields accessible through corresponding lenses are:

atResourceId :: Text

atTags :: [Tag]

Request lenses

atResourceId :: Lens' AddTags Text

The Amazon EMR resource identifier to which tags will be added. This value must be a cluster identifier.

atTags :: Lens' AddTags [Tag]

A list of tags to associate with a cluster and propagate to Amazon EC2 instances. Tags are user-defined key/value pairs that consist of a required key string with a maximum of 128 characters, and an optional value string with a maximum of 256 characters.

Response

data AddTagsResponse

Response constructor

addTagsResponse :: AddTagsResponse

AddTagsResponse constructor.