Network.AWS.EMR.DescribeCluster

Description

Provides cluster-level details including status, hardware and software configuration, VPC settings, and so on. For information about the cluster steps, see ListSteps.

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_DescribeCluster.html

Request

data DescribeCluster

Request constructor

describeCluster

Arguments

:: Text
dcClusterId

-> DescribeCluster

DescribeCluster constructor.

The fields accessible through corresponding lenses are:

dcClusterId :: Text

Request lenses

dcClusterId :: Lens' DescribeCluster Text

The identifier of the cluster to describe.

Response

data DescribeClusterResponse

Response constructor

describeClusterResponse

Arguments

:: Cluster
dcrCluster

-> DescribeClusterResponse

DescribeClusterResponse constructor.

The fields accessible through corresponding lenses are:

dcrCluster :: Cluster

Response lenses

dcrCluster :: Lens' DescribeClusterResponse Cluster

This output contains the details for the requested cluster.