Network.AWS.EMR.DescribeJobFlows

Description

This API is deprecated and will eventually be removed. We recommend you use ListClusters, DescribeCluster, ListSteps, ListInstanceGroups and ListBootstrapActions instead.

DescribeJobFlows returns a list of job flows that match all of the supplied parameters. The parameters can include a list of job flow IDs, job flow states, and restrictions on job flow creation date and time.

Regardless of supplied parameters, only job flows created within the last two months are returned.

If no parameters are supplied, then job flows matching either of the following criteria are returned:

Job flows created and completed in the last two weeks Job flows created within the last two months that are in one of the following states: RUNNING, WAITING, SHUTTING_DOWN, STARTING Amazon Elastic MapReduce can return a maximum of 512 job flow descriptions.

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_DescribeJobFlows.html

Request

data DescribeJobFlows

Request constructor

describeJobFlows :: DescribeJobFlows

DescribeJobFlows constructor.

The fields accessible through corresponding lenses are:

djfCreatedAfter :: Maybe UTCTime

djfCreatedBefore :: Maybe UTCTime

djfJobFlowIds :: [Text]

djfJobFlowStates :: [JobFlowExecutionState]

Request lenses

djfCreatedAfter :: Lens' DescribeJobFlows (Maybe UTCTime)

Return only job flows created after this date and time.

djfCreatedBefore :: Lens' DescribeJobFlows (Maybe UTCTime)

Return only job flows created before this date and time.

djfJobFlowIds :: Lens' DescribeJobFlows [Text]

Return only job flows whose job flow ID is contained in this list.

djfJobFlowStates :: Lens' DescribeJobFlows [JobFlowExecutionState]

Return only job flows whose state is contained in this list.

Response

data DescribeJobFlowsResponse

Response constructor

describeJobFlowsResponse :: DescribeJobFlowsResponse

DescribeJobFlowsResponse constructor.

The fields accessible through corresponding lenses are:

djfrJobFlows :: [JobFlowDetail]

Response lenses

djfrJobFlows :: Lens' DescribeJobFlowsResponse [JobFlowDetail]

A list of job flows matching the parameters supplied.