Network.AWS.EMR.DescribeStep

Description

Provides more detail about the cluster step.

Creating a Request

describeStep

Arguments

:: Text
dsClusterId

-> Text
dsStepId

-> DescribeStep

Creates a value of DescribeStep with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsClusterId - The identifier of the cluster with steps to describe.

dsStepId - The identifier of the step to describe.

data DescribeStep

This input determines which step to describe.

See: describeStep smart constructor.

Request Lenses

dsClusterId :: Lens' DescribeStep Text

The identifier of the cluster with steps to describe.

dsStepId :: Lens' DescribeStep Text

The identifier of the step to describe.

Destructuring the Response

describeStepResponse

Arguments

:: Int
dsrsResponseStatus

-> DescribeStepResponse

Creates a value of DescribeStepResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrsStep - The step details for the requested step identifier.

dsrsResponseStatus - -- | The response status code.

data DescribeStepResponse

This output contains the description of the cluster step.

See: describeStepResponse smart constructor.

Response Lenses

dsrsStep :: Lens' DescribeStepResponse (Maybe Step)

The step details for the requested step identifier.

dsrsResponseStatus :: Lens' DescribeStepResponse Int

- | The response status code.