Network.AWS.EMR.ListBootstrapActions

Description

Provides information about the bootstrap actions associated with a cluster.

This operation returns paginated results.

Creating a Request

listBootstrapActions

Arguments

:: Text
lbaClusterId

-> ListBootstrapActions

Creates a value of ListBootstrapActions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbaMarker - The pagination token that indicates the next set of results to retrieve.

lbaClusterId - The cluster identifier for the bootstrap actions to list.

data ListBootstrapActions

This input determines which bootstrap actions to retrieve.

See: listBootstrapActions smart constructor.

Request Lenses

lbaMarker :: Lens' ListBootstrapActions (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

lbaClusterId :: Lens' ListBootstrapActions Text

The cluster identifier for the bootstrap actions to list.

Destructuring the Response

listBootstrapActionsResponse

Arguments

:: Int
lbarsResponseStatus

-> ListBootstrapActionsResponse

Creates a value of ListBootstrapActionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbarsBootstrapActions - The bootstrap actions associated with the cluster.

lbarsMarker - The pagination token that indicates the next set of results to retrieve.

lbarsResponseStatus - -- | The response status code.

data ListBootstrapActionsResponse

This output contains the bootstrap actions detail.

See: listBootstrapActionsResponse smart constructor.

Response Lenses

lbarsBootstrapActions :: Lens' ListBootstrapActionsResponse [Command]

The bootstrap actions associated with the cluster.

lbarsMarker :: Lens' ListBootstrapActionsResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

lbarsResponseStatus :: Lens' ListBootstrapActionsResponse Int

- | The response status code.