Network.AWS.EMR.ListClusters

Description

Provides the status of all clusters visible to this AWS account. Allows you to filter the list of clusters based on certain criteria; for example, filtering by cluster creation date and time or by status. This call returns a maximum of 50 clusters per call, but returns a marker to track the paging of the cluster list across multiple ListClusters calls.

Request

data ListClusters

This input determines how the ListClusters action filters the list of clusters that it returns.

Request constructor

mkListClusters :: ListClusters

Smart constructor for the minimum required parameters to construct a valid ListClusters request.

The fields accessible through corresponding lenses are:

CreatedAfter :: Maybe POSIX

CreatedBefore :: Maybe POSIX

ClusterStates :: [ClusterState]

Marker :: Maybe Text

Request lenses

lcCreatedAfter :: Lens' ListClusters (Maybe POSIX)

The creation date and time beginning value filter for listing clusters .

lcCreatedBefore :: Lens' ListClusters (Maybe POSIX)

The creation date and time end value filter for listing clusters .

lcClusterStates :: Lens' ListClusters [ClusterState]

The cluster state filters to apply when listing clusters.

lcMarker :: Lens' ListClusters (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

Response

data ListClustersResponse

This contains a ClusterSummaryList with the cluster details; for example, the cluster IDs, names, and status.

Response constructor

mkListClustersResponse :: ListClustersResponse

Smart constructor for the minimum required parameters to construct a valid ListClustersResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Clusters :: [ClusterSummary]

Marker :: Maybe Text

Response lenses

lcrClusters :: Lens' ListClustersResponse [ClusterSummary]

The list of clusters for the account based on the given filters.

lcrMarker :: Lens' ListClustersResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.