Network.AWS.EMR.ListInstanceFleets

Description

Lists all available details about the instance fleets in a cluster.

This operation returns paginated results.

Creating a Request

listInstanceFleets

Arguments

:: Text
lifClusterId

-> ListInstanceFleets

Creates a value of ListInstanceFleets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lifMarker - The pagination token that indicates the next set of results to retrieve.

lifClusterId - The unique identifier of the cluster.

data ListInstanceFleets

See: listInstanceFleets smart constructor.

Request Lenses

lifMarker :: Lens' ListInstanceFleets (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

lifClusterId :: Lens' ListInstanceFleets Text

The unique identifier of the cluster.

Destructuring the Response

listInstanceFleetsResponse

Arguments

:: Int
lifrsResponseStatus

-> ListInstanceFleetsResponse

Creates a value of ListInstanceFleetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lifrsInstanceFleets - The list of instance fleets for the cluster and given filters.

lifrsMarker - The pagination token that indicates the next set of results to retrieve.

lifrsResponseStatus - -- | The response status code.

data ListInstanceFleetsResponse

See: listInstanceFleetsResponse smart constructor.

Response Lenses

lifrsInstanceFleets :: Lens' ListInstanceFleetsResponse [InstanceFleet]

The list of instance fleets for the cluster and given filters.

lifrsMarker :: Lens' ListInstanceFleetsResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

lifrsResponseStatus :: Lens' ListInstanceFleetsResponse Int

- | The response status code.