Network.AWS.EMR.ListInstances

Description

Provides information about the cluster instances that Amazon EMR provisions on behalf of a user when it creates the cluster. For example, this operation indicates when the EC2 instances reach the Ready state, when instances become available to Amazon EMR to use for jobs, and the IP addresses for cluster instances, etc.

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_ListInstances.html

Request

data ListInstances

Request constructor

listInstances

Arguments

:: Text
liClusterId

-> ListInstances

ListInstances constructor.

The fields accessible through corresponding lenses are:

liClusterId :: Text

liInstanceGroupId :: Maybe Text

liInstanceGroupTypes :: [InstanceGroupType]

liMarker :: Maybe Text

Request lenses

liClusterId :: Lens' ListInstances Text

The identifier of the cluster for which to list the instances.

liInstanceGroupId :: Lens' ListInstances (Maybe Text)

The identifier of the instance group for which to list the instances.

liInstanceGroupTypes :: Lens' ListInstances [InstanceGroupType]

The type of instance group for which to list the instances.

liMarker :: Lens' ListInstances (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

Response

data ListInstancesResponse

Response constructor

listInstancesResponse :: ListInstancesResponse

ListInstancesResponse constructor.

The fields accessible through corresponding lenses are:

lirInstances :: [Instance]

lirMarker :: Maybe Text

Response lenses

lirInstances :: Lens' ListInstancesResponse [Instance]

The list of instances for the cluster and given filters.

lirMarker :: Lens' ListInstancesResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.