Network.AWS.EMR.ListSteps

Description

Provides a list of steps for the cluster.

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_ListSteps.html

Request

data ListSteps

Request constructor

listSteps

Arguments

:: Text
lsClusterId

-> ListSteps

ListSteps constructor.

The fields accessible through corresponding lenses are:

lsClusterId :: Text

lsMarker :: Maybe Text

lsStepIds :: [Text]

lsStepStates :: [StepState]

Request lenses

lsClusterId :: Lens' ListSteps Text

The identifier of the cluster for which to list the steps.

lsMarker :: Lens' ListSteps (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

lsStepIds :: Lens' ListSteps [Text]

The filter to limit the step list based on the identifier of the steps.

lsStepStates :: Lens' ListSteps [StepState]

The filter to limit the step list based on certain states.

Response

data ListStepsResponse

Response constructor

listStepsResponse :: ListStepsResponse

ListStepsResponse constructor.

The fields accessible through corresponding lenses are:

lsrMarker :: Maybe Text

lsrSteps :: [StepSummary]

Response lenses

lsrMarker :: Lens' ListStepsResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

lsrSteps :: Lens' ListStepsResponse [StepSummary]

The filtered list of steps for the cluster.