Network.AWS.EMR.ModifyInstanceFleet

Description

Modifies the target On-Demand and target Spot capacities for the instance fleet with the specified InstanceFleetID within the cluster specified using ClusterID. The call either succeeds or fails atomically.

Creating a Request

modifyInstanceFleet

Arguments

:: Text
mifClusterId

-> InstanceFleetModifyConfig
mifInstanceFleet

-> ModifyInstanceFleet

Creates a value of ModifyInstanceFleet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mifClusterId - The unique identifier of the cluster.

mifInstanceFleet - The unique identifier of the instance fleet.

data ModifyInstanceFleet

See: modifyInstanceFleet smart constructor.

Request Lenses

mifClusterId :: Lens' ModifyInstanceFleet Text

The unique identifier of the cluster.

mifInstanceFleet :: Lens' ModifyInstanceFleet InstanceFleetModifyConfig

The unique identifier of the instance fleet.

Destructuring the Response

modifyInstanceFleetResponse :: ModifyInstanceFleetResponse

Creates a value of ModifyInstanceFleetResponse with the minimum fields required to make a request.

data ModifyInstanceFleetResponse

See: modifyInstanceFleetResponse smart constructor.