Network.AWS.EMR.ModifyInstanceGroups

Description

ModifyInstanceGroups modifies the number of nodes and configuration settings of an instance group. The input parameters include the new target instance count for the group and the instance group ID. The call will either succeed or fail atomically.

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_ModifyInstanceGroups.html

Request

data ModifyInstanceGroups

Request constructor

modifyInstanceGroups :: ModifyInstanceGroups

ModifyInstanceGroups constructor.

The fields accessible through corresponding lenses are:

migInstanceGroups :: [InstanceGroupModifyConfig]

Request lenses

migInstanceGroups :: Lens' ModifyInstanceGroups [InstanceGroupModifyConfig]

Instance groups to change.

Response

data ModifyInstanceGroupsResponse

Response constructor

modifyInstanceGroupsResponse :: ModifyInstanceGroupsResponse

ModifyInstanceGroupsResponse constructor.