Network.AWS.EMR.PutAutoScalingPolicy

Description

Creates or updates an automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. The automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric.

Creating a Request

putAutoScalingPolicy

Arguments

:: Text
paspClusterId

-> Text
paspInstanceGroupId

-> AutoScalingPolicy
paspAutoScalingPolicy

-> PutAutoScalingPolicy

Creates a value of PutAutoScalingPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

paspClusterId - Specifies the ID of a cluster. The instance group to which the automatic scaling policy is applied is within this cluster.

paspInstanceGroupId - Specifies the ID of the instance group to which the automatic scaling policy is applied.

paspAutoScalingPolicy - Specifies the definition of the automatic scaling policy.

data PutAutoScalingPolicy

See: putAutoScalingPolicy smart constructor.

Request Lenses

paspClusterId :: Lens' PutAutoScalingPolicy Text

Specifies the ID of a cluster. The instance group to which the automatic scaling policy is applied is within this cluster.

paspInstanceGroupId :: Lens' PutAutoScalingPolicy Text

Specifies the ID of the instance group to which the automatic scaling policy is applied.

paspAutoScalingPolicy :: Lens' PutAutoScalingPolicy AutoScalingPolicy

Specifies the definition of the automatic scaling policy.

Destructuring the Response

putAutoScalingPolicyResponse

Arguments

:: Int
pasprsResponseStatus

-> PutAutoScalingPolicyResponse

Creates a value of PutAutoScalingPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pasprsClusterId - Specifies the ID of a cluster. The instance group to which the automatic scaling policy is applied is within this cluster.

pasprsAutoScalingPolicy - The automatic scaling policy definition.

pasprsInstanceGroupId - Specifies the ID of the instance group to which the scaling policy is applied.

pasprsResponseStatus - -- | The response status code.

data PutAutoScalingPolicyResponse

See: putAutoScalingPolicyResponse smart constructor.

Response Lenses

pasprsClusterId :: Lens' PutAutoScalingPolicyResponse (Maybe Text)

Specifies the ID of a cluster. The instance group to which the automatic scaling policy is applied is within this cluster.

pasprsAutoScalingPolicy :: Lens' PutAutoScalingPolicyResponse (Maybe AutoScalingPolicyDescription)

The automatic scaling policy definition.

pasprsInstanceGroupId :: Lens' PutAutoScalingPolicyResponse (Maybe Text)

Specifies the ID of the instance group to which the scaling policy is applied.

pasprsResponseStatus :: Lens' PutAutoScalingPolicyResponse Int

- | The response status code.