Network.AWS.EMR.RemoveTags

Description

Removes tags from an Amazon EMR resource. Tags make it easier to associate clusters in various ways, such as grouping clusters to track your Amazon EMR resource allocation costs. For more information, see Tagging Amazon EMRResources.

The following example removes the stack tag with value Prod from a cluster:

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_RemoveTags.html

Request

data RemoveTags

Request constructor

removeTags

Arguments

:: Text
rtResourceId

-> RemoveTags

RemoveTags constructor.

The fields accessible through corresponding lenses are:

rtResourceId :: Text

rtTagKeys :: [Text]

Request lenses

rtResourceId :: Lens' RemoveTags Text

The Amazon EMR resource identifier from which tags will be removed. This value must be a cluster identifier.

rtTagKeys :: Lens' RemoveTags [Text]

A list of tag keys to remove from a resource.

Response

data RemoveTagsResponse

Response constructor

removeTagsResponse :: RemoveTagsResponse

RemoveTagsResponse constructor.