Network.AWS.EMR.SetTerminationProtection

Description

SetTerminationProtection locks a job flow so the Amazon EC2 instances in the cluster cannot be terminated by user intervention, an API call, or in the event of a job-flow error. The cluster still terminates upon successful completion of the job flow. Calling SetTerminationProtection on a job flow is analogous to calling the Amazon EC2 DisableAPITermination API on all of the EC2 instances in a cluster.

SetTerminationProtection is used to prevent accidental termination of a job flow and to ensure that in the event of an error, the instances will persist so you can recover any data stored in their ephemeral instance storage.

To terminate a job flow that has been locked by setting SetTerminationProtection to true, you must first unlock the job flow by a subsequent call to SetTerminationProtection in which you set the value to false.

For more information, go to Protecting a Job Flow from Termination in the Amazon Elastic MapReduce Developer's Guide.

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_SetTerminationProtection.html

Request

data SetTerminationProtection

Request constructor

setTerminationProtection

Arguments

:: Bool
stpTerminationProtected

-> SetTerminationProtection

SetTerminationProtection constructor.

The fields accessible through corresponding lenses are:

stpJobFlowIds :: [Text]

stpTerminationProtected :: Bool

Request lenses

stpJobFlowIds :: Lens' SetTerminationProtection [Text]

A list of strings that uniquely identify the job flows to protect. This identifier is returned by RunJobFlow and can also be obtained from DescribeJobFlows .

stpTerminationProtected :: Lens' SetTerminationProtection Bool

A Boolean that indicates whether to protect the job flow and prevent the Amazon EC2 instances in the cluster from shutting down due to API calls, user intervention, or job-flow error.

Response

data SetTerminationProtectionResponse

Response constructor

setTerminationProtectionResponse :: SetTerminationProtectionResponse

SetTerminationProtectionResponse constructor.