Network.AWS.EMR.Types

Description

Amazon Elastic MapReduce (Amazon EMR) is a web service that makes it easy to process large amounts of data efficiently. Amazon EMR uses Hadoop processing combined with several AWS products to do such tasks as web indexing, data mining, log file analysis, machine learning, scientific simulation, and data warehousing.

Service

data EMR

Supported version (2009-03-31) of the Amazon Elastic MapReduce service.

Errors

data EMRError

A sum type representing possible errors returned by the EMR service.

These typically include HTTPExceptions thrown by the underlying HTTP mechanisms, serialisation errors, and typed errors as specified by the service description where applicable.

Constructors

EMRClient HttpException

EMRSerializer String

EMRService String

InternalServerError
Indicates that an error occurred while processing the request and that the request was not completed.

InternalServerException
This exception occurs when there is an internal failure in the EMR service.

Fields

_ise1Message :: Maybe Text

InvalidRequestException
This exception occurs when there is something wrong with user input.

Fields

_ireErrorCode :: Maybe Text

_ireMessage :: Maybe Text

_EMRClient :: Prism' EMRError HttpException

See: EMRClient

_EMRSerializer :: Prism' EMRError String

See: EMRSerializer

_EMRService :: Prism' EMRError String

See: EMRService

_InternalServerError :: Prism' EMRError ()

Indicates that an error occurred while processing the request and that the request was not completed.

See: InternalServerError

_InternalServerException :: Prism' EMRError (Maybe Text)

This exception occurs when there is an internal failure in the EMR service.

See: InternalServerException

_InvalidRequestException :: Prism' EMRError (Maybe Text, Maybe Text)

This exception occurs when there is something wrong with user input.

See: InvalidRequestException

ActionOnFailure

data ActionOnFailure

Constructors

ActionOnFailureCancelAndWait
CANCEL_AND_WAIT

ActionOnFailureContinue
CONTINUE

ActionOnFailureTerminateCluster
TERMINATE_CLUSTER

ActionOnFailureTerminateJobFlow
TERMINATE_JOB_FLOW

ClusterState

data ClusterState

Constructors

ClusterStateBootstrapping
BOOTSTRAPPING

ClusterStateRunning
RUNNING

ClusterStateStarting
STARTING

ClusterStateTerminated
TERMINATED

ClusterStateTerminatedWithErrors
TERMINATED_WITH_ERRORS

ClusterStateTerminating
TERMINATING

ClusterStateWaiting
WAITING

ClusterStateChangeReasonCode

data ClusterStateChangeReasonCode

Constructors

ClusterStateChangeReasonCodeAllStepsCompleted
ALL_STEPS_COMPLETED

ClusterStateChangeReasonCodeBootstrapFailure
BOOTSTRAP_FAILURE

ClusterStateChangeReasonCodeInstanceFailure
INSTANCE_FAILURE

ClusterStateChangeReasonCodeInternalError
INTERNAL_ERROR

ClusterStateChangeReasonCodeStepFailure
STEP_FAILURE

ClusterStateChangeReasonCodeUserRequest
USER_REQUEST

ClusterStateChangeReasonCodeValidationError
VALIDATION_ERROR

InstanceGroupState

data InstanceGroupState

Constructors

InstanceGroupStateArrested
ARRESTED

InstanceGroupStateBootstrapping
BOOTSTRAPPING

InstanceGroupStateEnded
ENDED

InstanceGroupStateProvisioning
PROVISIONING

InstanceGroupStateResizing
RESIZING

InstanceGroupStateRunning
RUNNING

InstanceGroupStateShuttingDown
SHUTTING_DOWN

InstanceGroupStateSuspended
SUSPENDED

InstanceGroupStateTerminated
TERMINATED

InstanceGroupStateTerminating
TERMINATING

InstanceGroupStateChangeReasonCode

data InstanceGroupStateChangeReasonCode

Constructors

InstanceGroupStateChangeReasonCodeClusterTerminated
CLUSTER_TERMINATED

InstanceGroupStateChangeReasonCodeInstanceFailure
INSTANCE_FAILURE

InstanceGroupStateChangeReasonCodeInternalError
INTERNAL_ERROR

InstanceGroupStateChangeReasonCodeValidationError
VALIDATION_ERROR

InstanceGroupType

data InstanceGroupType

Constructors

InstanceGroupTypeCore
CORE

InstanceGroupTypeMaster
MASTER

InstanceGroupTypeTask
TASK

InstanceRoleType

data InstanceRoleType

Constructors

InstanceRoleTypeCore
CORE

InstanceRoleTypeMaster
MASTER

InstanceRoleTypeTask
TASK

InstanceState

data InstanceState

Constructors

InstanceStateAwaitingFulfillment
AWAITING_FULFILLMENT

InstanceStateBootstrapping
BOOTSTRAPPING

InstanceStateProvisioning
PROVISIONING

InstanceStateRunning
RUNNING

InstanceStateTerminated
TERMINATED

InstanceStateChangeReasonCode

data InstanceStateChangeReasonCode

Constructors

InstanceStateChangeReasonCodeBootstrapFailure
BOOTSTRAP_FAILURE

InstanceStateChangeReasonCodeClusterTerminated
CLUSTER_TERMINATED

InstanceStateChangeReasonCodeInstanceFailure
INSTANCE_FAILURE

InstanceStateChangeReasonCodeInternalError
INTERNAL_ERROR

InstanceStateChangeReasonCodeValidationError
VALIDATION_ERROR

JobFlowExecutionState

data JobFlowExecutionState

Constructors

JobFlowExecutionStateBootstrapping
BOOTSTRAPPING

JobFlowExecutionStateCompleted
COMPLETED

JobFlowExecutionStateFailed
FAILED

JobFlowExecutionStateRunning
RUNNING

JobFlowExecutionStateShuttingDown
SHUTTING_DOWN

JobFlowExecutionStateStarting
STARTING

JobFlowExecutionStateTerminated
TERMINATED

JobFlowExecutionStateWaiting
WAITING

MarketType

data MarketType

Constructors

MarketTypeOnDemand
ON_DEMAND

MarketTypeSpot
SPOT

StepExecutionState

data StepExecutionState

Constructors

StepExecutionStateCancelled
CANCELLED

StepExecutionStateCompleted
COMPLETED

StepExecutionStateContinue
CONTINUE

StepExecutionStateFailed
FAILED

StepExecutionStateInterrupted
INTERRUPTED

StepExecutionStatePending
PENDING

StepExecutionStateRunning
RUNNING

StepState

data StepState

Constructors

StepStateCancelled
CANCELLED

StepStateCompleted
COMPLETED

StepStateFailed
FAILED

StepStateInterrupted
INTERRUPTED

StepStatePending
PENDING

StepStateRunning
RUNNING

StepStateChangeReasonCode

data StepStateChangeReasonCode

Constructors

StepStateChangeReasonCodeNone
NONE

BootstrapActionDetail

data BootstrapActionDetail

Reports the configuration of a bootstrap action in a job flow.

bootstrapActionDetail :: BootstrapActionDetail

Smart constructor for the minimum required fields to construct a valid BootstrapActionDetail data type to populate a request.

The fields accessible through corresponding lenses are:

BootstrapActionConfig :: Maybe BootstrapActionConfig

badBootstrapActionConfig :: Lens' BootstrapActionDetail (Maybe BootstrapActionConfig)

A description of the bootstrap action.

PlacementType

data PlacementType

The Amazon EC2 Availability Zone for the job flow.

placementType

Arguments

:: Text
ptAvailabilityZone

-> PlacementType

Smart constructor for the minimum required fields to construct a valid PlacementType data type to populate a request.

The fields accessible through corresponding lenses are:

AvailabilityZone :: Text

ptAvailabilityZone :: Lens' PlacementType Text

The Amazon EC2 Availability Zone for the job flow.

Application

data Application

An application is any Amazon or third-party software that you can add to the cluster. This structure contains a list of strings that indicates the software to use with the cluster and accepts a user argument list. Amazon EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action argument. For more information, see Launch a Job Flow on the MapR Distribution for Hadoop. Currently supported values are: "mapr-m3" - launch the job flow using MapR M3 Edition. "mapr-m5" - launch the job flow using MapR M5 Edition. "mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition, respectively.

application :: Application

Smart constructor for the minimum required fields to construct a valid Application data type.

Application is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Name :: Maybe Text

Version :: Maybe Text

Args :: [Text]

AdditionalInfo :: Map Text Text

aName :: Lens' Application (Maybe Text)

The name of the application.

aVersion :: Lens' Application (Maybe Text)

The version of the application.

aArgs :: Lens' Application [Text]

Arguments for Amazon EMR to pass to the application.

aAdditionalInfo :: Lens' Application (Map Text Text)

This option is for advanced users only. This is meta information about third-party applications that third-party vendors use for testing purposes.

BootstrapActionConfig

data BootstrapActionConfig

A description of the bootstrap action.

bootstrapActionConfig

Arguments

:: Text
bacName

-> ScriptBootstrapActionConfig
bacScriptBootstrapAction

-> BootstrapActionConfig

Smart constructor for the minimum required fields to construct a valid BootstrapActionConfig data type to populate a request.

The fields accessible through corresponding lenses are:

Name :: Text

ScriptBootstrapAction :: ScriptBootstrapActionConfig

bacName :: Lens' BootstrapActionConfig Text

The name of the bootstrap action.

bacScriptBootstrapAction :: Lens' BootstrapActionConfig ScriptBootstrapActionConfig

The script run by the bootstrap action.

Cluster

data Cluster

This output contains the details for the requested cluster.

cluster :: Cluster

Smart constructor for the minimum required fields to construct a valid Cluster data type.

Cluster is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Id :: Maybe Text

Name :: Maybe Text

Status :: Maybe ClusterStatus

Ec2InstanceAttributes :: Maybe Ec2InstanceAttributes

LogUri :: Maybe Text

RequestedAmiVersion :: Maybe Text

RunningAmiVersion :: Maybe Text

AutoTerminate :: Maybe Bool

TerminationProtected :: Maybe Bool

VisibleToAllUsers :: Maybe Bool

Applications :: [Application]

Tags :: [Tag]

ServiceRole :: Maybe Text

cId :: Lens' Cluster (Maybe Text)

The unique identifier for the cluster.

cName :: Lens' Cluster (Maybe Text)

The name of the cluster.

cStatus :: Lens' Cluster (Maybe ClusterStatus)

The current status details about the cluster.

cEc2InstanceAttributes :: Lens' Cluster (Maybe Ec2InstanceAttributes)

Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

cLogUri :: Lens' Cluster (Maybe Text)

The path to the Amazon S3 location where logs for this cluster are stored.

cRequestedAmiVersion :: Lens' Cluster (Maybe Text)

The AMI version requested for this cluster.JobFlowDetail$AmiVersion.-->.

cRunningAmiVersion :: Lens' Cluster (Maybe Text)

The AMI version running on this cluster. This differs from the requested version only if the requested version is a meta version, such as "latest". JobFlowDetail$AmiVersion.-->.

cAutoTerminate :: Lens' Cluster (Maybe Bool)

Specifies whether the cluster should terminate after completing all steps.

cTerminationProtected :: Lens' Cluster (Maybe Bool)

Indicates whether Amazon EMR will lock the cluster to prevent the EC2 instances from being terminated by an API call or user intervention, or in the event of a cluster error.

cVisibleToAllUsers :: Lens' Cluster (Maybe Bool)

Indicates whether the job flow is visible to all IAM users of the AWS account associated with the job flow. If this value is set to true, all IAM users of that AWS account can view and manage the job flow if they have the proper policy permissions set. If this value is false, only the IAM user that created the cluster can view and manage it. This value can be changed using the SetVisibleToAllUsers action.

cApplications :: Lens' Cluster [Application]

The applications installed on this cluster.

cTags :: Lens' Cluster [Tag]

A list of tags associated with a cluster.

cServiceRole :: Lens' Cluster (Maybe Text)

The IAM role that will be assumed by the Amazon EMR service to access AWS resources on your behalf.

ClusterStateChangeReason

data ClusterStateChangeReason

The reason for the cluster status change.

clusterStateChangeReason :: ClusterStateChangeReason

Smart constructor for the minimum required fields to construct a valid ClusterStateChangeReason data type to populate a request.

The fields accessible through corresponding lenses are:

Code :: Maybe ClusterStateChangeReasonCode

Message :: Maybe Text

cscrCode :: Lens' ClusterStateChangeReason (Maybe ClusterStateChangeReasonCode)

The programmatic code for the state change reason.

cscrMessage :: Lens' ClusterStateChangeReason (Maybe Text)

The descriptive message for the state change reason.

ClusterStatus

data ClusterStatus

The current status details about the cluster.

clusterStatus :: ClusterStatus

Smart constructor for the minimum required fields to construct a valid ClusterStatus data type to populate a request.

The fields accessible through corresponding lenses are:

State :: Maybe ClusterState

StateChangeReason :: Maybe ClusterStateChangeReason

Timeline :: Maybe ClusterTimeline

csState :: Lens' ClusterStatus (Maybe ClusterState)

The current state of the cluster.

csStateChangeReason :: Lens' ClusterStatus (Maybe ClusterStateChangeReason)

The reason for the cluster status change.

csTimeline :: Lens' ClusterStatus (Maybe ClusterTimeline)

A timeline that represents the status of a cluster over the lifetime of the cluster.

ClusterSummary

data ClusterSummary

The summary description of the cluster.

clusterSummary :: ClusterSummary

Smart constructor for the minimum required fields to construct a valid ClusterSummary data type.

ClusterSummary is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Id :: Maybe Text

Name :: Maybe Text

Status :: Maybe ClusterStatus

csrId :: Lens' ClusterSummary (Maybe Text)

The unique identifier for the cluster.

csrName :: Lens' ClusterSummary (Maybe Text)

The name of the cluster.

csrStatus :: Lens' ClusterSummary (Maybe ClusterStatus)

The details about the current status of the cluster.

ClusterTimeline

data ClusterTimeline

A timeline that represents the status of a cluster over the lifetime of the cluster.

clusterTimeline :: ClusterTimeline

Smart constructor for the minimum required fields to construct a valid ClusterTimeline data type to populate a request.

The fields accessible through corresponding lenses are:

CreationDateTime :: Maybe POSIX

ReadyDateTime :: Maybe POSIX

EndDateTime :: Maybe POSIX

ctCreationDateTime :: Lens' ClusterTimeline (Maybe POSIX)

The creation date and time of the cluster.

ctReadyDateTime :: Lens' ClusterTimeline (Maybe POSIX)

The date and time when the cluster was ready to execute steps.

ctEndDateTime :: Lens' ClusterTimeline (Maybe POSIX)

The date and time when the cluster was terminated.

Command

data Command

An entity describing an executable that runs on a cluster.

command :: Command

Smart constructor for the minimum required fields to construct a valid Command data type.

Command is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Name :: Maybe Text

ScriptPath :: Maybe Text

Args :: [Text]

crName :: Lens' Command (Maybe Text)

The name of the command.

crScriptPath :: Lens' Command (Maybe Text)

The Amazon S3 location of the command script.

crArgs :: Lens' Command [Text]

Arguments for Amazon EMR to pass to the command for execution.

Ec2InstanceAttributes

data Ec2InstanceAttributes

Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

ec2InstanceAttributes :: Ec2InstanceAttributes

Smart constructor for the minimum required fields to construct a valid Ec2InstanceAttributes data type.

Ec2InstanceAttributes is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Ec2KeyName :: Maybe Text

Ec2SubnetId :: Maybe Text

Ec2AvailabilityZone :: Maybe Text

IamInstanceProfile :: Maybe Text

eiaEc2KeyName :: Lens' Ec2InstanceAttributes (Maybe Text)

The name of the Amazon EC2 key pair to use when connecting with SSH into the master node as a user named "hadoop".

eiaEc2SubnetId :: Lens' Ec2InstanceAttributes (Maybe Text)

To launch the job flow in Amazon VPC, set this parameter to the identifier of the Amazon VPC subnet where you want the job flow to launch. If you do not specify this value, the job flow is launched in the normal AWS cloud, outside of a VPC. Amazon VPC currently does not support cluster compute quadruple extra large (cc1.4xlarge) instances. Thus, you cannot specify the cc1.4xlarge instance type for nodes of a job flow launched in a VPC.

eiaEc2AvailabilityZone :: Lens' Ec2InstanceAttributes (Maybe Text)

The Availability Zone in which the cluster will run.

eiaIamInstanceProfile :: Lens' Ec2InstanceAttributes (Maybe Text)

The IAM role that was specified when the job flow was launched. The EC2 instances of the job flow assume this role.

HadoopJarStepConfig

data HadoopJarStepConfig

The JAR file used for the job flow step.

hadoopJarStepConfig

Arguments

:: Text
hjscJar

-> HadoopJarStepConfig

Smart constructor for the minimum required fields to construct a valid HadoopJarStepConfig data type to populate a request.

The fields accessible through corresponding lenses are:

Properties :: [KeyValue]

Jar :: Text

MainClass :: Maybe Text

Args :: [Text]

hjscProperties :: Lens' HadoopJarStepConfig [KeyValue]

A list of Java properties that are set when the step runs. You can use these properties to pass key value pairs to your main function.

hjscJar :: Lens' HadoopJarStepConfig Text

A path to a JAR file run during the step.

hjscMainClass :: Lens' HadoopJarStepConfig (Maybe Text)

The name of the main class in the specified Java file. If not specified, the JAR file should specify a Main-Class in its manifest file.

hjscArgs :: Lens' HadoopJarStepConfig [Text]

A list of command line arguments passed to the JAR file's main function when executed.

HadoopStepConfig

data HadoopStepConfig

The Hadoop job configuration of the cluster step.

hadoopStepConfig :: HadoopStepConfig

Smart constructor for the minimum required fields to construct a valid HadoopStepConfig data type.

HadoopStepConfig is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Jar :: Maybe Text

Properties :: Map Text Text

MainClass :: Maybe Text

Args :: [Text]

hscJar :: Lens' HadoopStepConfig (Maybe Text)

The path to the JAR file that runs during the step.

hscProperties :: Lens' HadoopStepConfig (Map Text Text)

The list of Java properties that are set when the step runs. You can use these properties to pass key value pairs to your main function.

hscMainClass :: Lens' HadoopStepConfig (Maybe Text)

The name of the main class in the specified Java file. If not specified, the JAR file should specify a main class in its manifest file.

hscArgs :: Lens' HadoopStepConfig [Text]

The list of command line arguments to pass to the JAR file's main function for execution.

Instance

data Instance

Represents an EC2 instance provisioned as part of cluster.

instance' :: Instance

Smart constructor for the minimum required fields to construct a valid Instance data type.

Instance is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Id :: Maybe Text

Ec2InstanceId :: Maybe Text

PublicDnsName :: Maybe Text

PublicIpAddress :: Maybe Text

PrivateDnsName :: Maybe Text

PrivateIpAddress :: Maybe Text

Status :: Maybe InstanceStatus

iId :: Lens' Instance (Maybe Text)

The unique identifier for the instance in Amazon EMR.

iEc2InstanceId :: Lens' Instance (Maybe Text)

The unique identifier of the instance in Amazon EC2.

iPublicDnsName :: Lens' Instance (Maybe Text)

The public DNS name of the instance.

iPublicIpAddress :: Lens' Instance (Maybe Text)

The public IP address of the instance.

iPrivateDnsName :: Lens' Instance (Maybe Text)

The private DNS name of the instance.

iPrivateIpAddress :: Lens' Instance (Maybe Text)

The private IP address of the instance.

iStatus :: Lens' Instance (Maybe InstanceStatus)

The current status of the instance.

InstanceGroup

data InstanceGroup

This entity represents an instance group, which is a group of instances that have common purpose. For example, CORE instance group is used for HDFS.

instanceGroup :: InstanceGroup

Smart constructor for the minimum required fields to construct a valid InstanceGroup data type.

InstanceGroup is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Id :: Maybe Text

Name :: Maybe Text

Market :: Maybe MarketType

InstanceGroupType :: Maybe InstanceGroupType

BidPrice :: Maybe Text

InstanceType :: Maybe Text

RequestedInstanceCount :: Maybe Integer

RunningInstanceCount :: Maybe Integer

Status :: Maybe InstanceGroupStatus

igId :: Lens' InstanceGroup (Maybe Text)

The identifier of the instance group.

igName :: Lens' InstanceGroup (Maybe Text)

The name of the instance group.

igMarket :: Lens' InstanceGroup (Maybe MarketType)

The marketplace to provision instances for this group. Valid values are ON_DEMAND or SPOT.

igInstanceGroupType :: Lens' InstanceGroup (Maybe InstanceGroupType)

The type of the instance group. Valid values are MASTER, CORE or TASK.

igBidPrice :: Lens' InstanceGroup (Maybe Text)

The bid price for each EC2 instance in the instance group when launching nodes as Spot Instances, expressed in USD.

igInstanceType :: Lens' InstanceGroup (Maybe Text)

The EC2 instance type for all instances in the instance group.

igRequestedInstanceCount :: Lens' InstanceGroup (Maybe Integer)

The target number of instances for the instance group.

igRunningInstanceCount :: Lens' InstanceGroup (Maybe Integer)

The number of instances currently running in this instance group.

igStatus :: Lens' InstanceGroup (Maybe InstanceGroupStatus)

The current status of the instance group.

InstanceGroupConfig

data InstanceGroupConfig

Configuration defining a new instance group.

instanceGroupConfig

Arguments

:: InstanceRoleType
igcInstanceRole

-> Text
igcInstanceType

-> Integer
igcInstanceCount

-> InstanceGroupConfig

Smart constructor for the minimum required fields to construct a valid InstanceGroupConfig data type to populate a request.

The fields accessible through corresponding lenses are:

Name :: Maybe Text

Market :: Maybe MarketType

InstanceRole :: InstanceRoleType

BidPrice :: Maybe Text

InstanceType :: Text

InstanceCount :: Integer

igcName :: Lens' InstanceGroupConfig (Maybe Text)

Friendly name given to the instance group.

igcMarket :: Lens' InstanceGroupConfig (Maybe MarketType)

Market type of the Amazon EC2 instances used to create a cluster node.

igcInstanceRole :: Lens' InstanceGroupConfig InstanceRoleType

The role of the instance group in the cluster.

igcBidPrice :: Lens' InstanceGroupConfig (Maybe Text)

Bid price for each Amazon EC2 instance in the instance group when launching nodes as Spot Instances, expressed in USD.

igcInstanceType :: Lens' InstanceGroupConfig Text

The Amazon EC2 instance type for all instances in the instance group.

igcInstanceCount :: Lens' InstanceGroupConfig Integer

Target number of instances for the instance group.

InstanceGroupDetail

data InstanceGroupDetail

Detailed information about an instance group.

instanceGroupDetail

Arguments

:: POSIX
igdCreationDateTime

-> MarketType
igdMarket

-> InstanceRoleType
igdInstanceRole

-> Text
igdInstanceType

-> Integer
igdInstanceRequestCount

-> Integer
igdInstanceRunningCount

-> InstanceGroupState
igdState

-> InstanceGroupDetail

Smart constructor for the minimum required fields to construct a valid InstanceGroupDetail data type to populate a request.

The fields accessible through corresponding lenses are:

InstanceGroupId :: Maybe Text

Name :: Maybe Text

Market :: MarketType

InstanceRole :: InstanceRoleType

BidPrice :: Maybe Text

InstanceType :: Text

InstanceRequestCount :: Integer

InstanceRunningCount :: Integer

State :: InstanceGroupState

LastStateChangeReason :: Maybe Text

CreationDateTime :: POSIX

StartDateTime :: Maybe POSIX

ReadyDateTime :: Maybe POSIX

EndDateTime :: Maybe POSIX

igdInstanceGroupId :: Lens' InstanceGroupDetail (Maybe Text)

Unique identifier for the instance group.

igdName :: Lens' InstanceGroupDetail (Maybe Text)

Friendly name for the instance group.

igdMarket :: Lens' InstanceGroupDetail MarketType

Market type of the Amazon EC2 instances used to create a cluster node.

igdInstanceRole :: Lens' InstanceGroupDetail InstanceRoleType

Instance group role in the cluster.

igdBidPrice :: Lens' InstanceGroupDetail (Maybe Text)

Bid price for EC2 Instances when launching nodes as Spot Instances, expressed in USD.

igdInstanceType :: Lens' InstanceGroupDetail Text

Amazon EC2 Instance type.

igdInstanceRequestCount :: Lens' InstanceGroupDetail Integer

Target number of instances to run in the instance group.

igdInstanceRunningCount :: Lens' InstanceGroupDetail Integer

Actual count of running instances.

igdState :: Lens' InstanceGroupDetail InstanceGroupState

State of instance group. The following values are deprecated: STARTING, TERMINATED, and FAILED.

igdLastStateChangeReason :: Lens' InstanceGroupDetail (Maybe Text)

Details regarding the state of the instance group.

igdCreationDateTime :: Lens' InstanceGroupDetail POSIX

The date/time the instance group was created.

igdStartDateTime :: Lens' InstanceGroupDetail (Maybe POSIX)

The date/time the instance group was started.

igdReadyDateTime :: Lens' InstanceGroupDetail (Maybe POSIX)

The date/time the instance group was available to the cluster.

igdEndDateTime :: Lens' InstanceGroupDetail (Maybe POSIX)

The date/time the instance group was terminated.

InstanceGroupModifyConfig

data InstanceGroupModifyConfig

Modify an instance group size.

instanceGroupModifyConfig

Arguments

:: Text
igmcInstanceGroupId

-> InstanceGroupModifyConfig

Smart constructor for the minimum required fields to construct a valid InstanceGroupModifyConfig data type to populate a request.

The fields accessible through corresponding lenses are:

InstanceGroupId :: Text

InstanceCount :: Maybe Integer

EC2InstanceIdsToTerminate :: [Text]

igmcInstanceGroupId :: Lens' InstanceGroupModifyConfig Text

Unique ID of the instance group to expand or shrink.

igmcInstanceCount :: Lens' InstanceGroupModifyConfig (Maybe Integer)

Target size for the instance group.

igmcEC2InstanceIdsToTerminate :: Lens' InstanceGroupModifyConfig [Text]

The EC2 InstanceIds to terminate. For advanced users only. Once you terminate the instances, the instance group will not return to its original requested size.

InstanceGroupStateChangeReason

data InstanceGroupStateChangeReason

The status change reason details for the instance group.

instanceGroupStateChangeReason :: InstanceGroupStateChangeReason

Smart constructor for the minimum required fields to construct a valid InstanceGroupStateChangeReason data type to populate a request.

The fields accessible through corresponding lenses are:

Code :: Maybe InstanceGroupStateChangeReasonCode

Message :: Maybe Text

igscrCode :: Lens' InstanceGroupStateChangeReason (Maybe InstanceGroupStateChangeReasonCode)

The programmable code for the state change reason.

igscrMessage :: Lens' InstanceGroupStateChangeReason (Maybe Text)

The status change reason description.

InstanceGroupStatus

data InstanceGroupStatus

The current status of the instance group.

instanceGroupStatus :: InstanceGroupStatus

Smart constructor for the minimum required fields to construct a valid InstanceGroupStatus data type to populate a request.

The fields accessible through corresponding lenses are:

State :: Maybe InstanceGroupState

StateChangeReason :: Maybe InstanceGroupStateChangeReason

Timeline :: Maybe InstanceGroupTimeline

igsState :: Lens' InstanceGroupStatus (Maybe InstanceGroupState)

The current state of the instance group.

igsStateChangeReason :: Lens' InstanceGroupStatus (Maybe InstanceGroupStateChangeReason)

The status change reason details for the instance group.

igsTimeline :: Lens' InstanceGroupStatus (Maybe InstanceGroupTimeline)

The timeline of the instance group status over time.

InstanceGroupTimeline

data InstanceGroupTimeline

The timeline of the instance group status over time.

instanceGroupTimeline :: InstanceGroupTimeline

Smart constructor for the minimum required fields to construct a valid InstanceGroupTimeline data type to populate a request.

The fields accessible through corresponding lenses are:

CreationDateTime :: Maybe POSIX

ReadyDateTime :: Maybe POSIX

EndDateTime :: Maybe POSIX

igtCreationDateTime :: Lens' InstanceGroupTimeline (Maybe POSIX)

The creation date and time of the instance group.

igtReadyDateTime :: Lens' InstanceGroupTimeline (Maybe POSIX)

The date and time when the instance group became ready to perform tasks.

igtEndDateTime :: Lens' InstanceGroupTimeline (Maybe POSIX)

The date and time when the instance group terminated.

InstanceStateChangeReason

data InstanceStateChangeReason

The details of the status change reason for the instance.

instanceStateChangeReason :: InstanceStateChangeReason

Smart constructor for the minimum required fields to construct a valid InstanceStateChangeReason data type to populate a request.

The fields accessible through corresponding lenses are:

Code :: Maybe InstanceStateChangeReasonCode

Message :: Maybe Text

iscrCode :: Lens' InstanceStateChangeReason (Maybe InstanceStateChangeReasonCode)

The programmable code for the state change reason.

iscrMessage :: Lens' InstanceStateChangeReason (Maybe Text)

The status change reason description.

InstanceStatus

data InstanceStatus

The current status of the instance.

instanceStatus :: InstanceStatus

Smart constructor for the minimum required fields to construct a valid InstanceStatus data type to populate a request.

The fields accessible through corresponding lenses are:

State :: Maybe InstanceState

StateChangeReason :: Maybe InstanceStateChangeReason

Timeline :: Maybe InstanceTimeline

isState :: Lens' InstanceStatus (Maybe InstanceState)

The current state of the instance.

isStateChangeReason :: Lens' InstanceStatus (Maybe InstanceStateChangeReason)

The details of the status change reason for the instance.

isTimeline :: Lens' InstanceStatus (Maybe InstanceTimeline)

The timeline of the instance status over time.

InstanceTimeline

data InstanceTimeline

The timeline of the instance status over time.

instanceTimeline :: InstanceTimeline

Smart constructor for the minimum required fields to construct a valid InstanceTimeline data type to populate a request.

The fields accessible through corresponding lenses are:

CreationDateTime :: Maybe POSIX

ReadyDateTime :: Maybe POSIX

EndDateTime :: Maybe POSIX

itCreationDateTime :: Lens' InstanceTimeline (Maybe POSIX)

The creation date and time of the instance.

itReadyDateTime :: Lens' InstanceTimeline (Maybe POSIX)

The date and time when the instance was ready to perform tasks.

itEndDateTime :: Lens' InstanceTimeline (Maybe POSIX)

The date and time when the instance was terminated.

JobFlowDetail

data JobFlowDetail

A description of a job flow.

jobFlowDetail

Arguments

:: Text
jfdJobFlowId

-> Text
jfdName

-> JobFlowExecutionStatusDetail
jfdExecutionStatusDetail

-> JobFlowInstancesDetail
jfdInstances

-> JobFlowDetail

Smart constructor for the minimum required fields to construct a valid JobFlowDetail data type.

JobFlowDetail is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

JobFlowId :: Text

Name :: Text

LogUri :: Maybe Text

AmiVersion :: Maybe Text

ExecutionStatusDetail :: JobFlowExecutionStatusDetail

Instances :: JobFlowInstancesDetail

Steps :: [StepDetail]

BootstrapActions :: [BootstrapActionDetail]

SupportedProducts :: [Text]

VisibleToAllUsers :: Maybe Bool

JobFlowRole :: Maybe Text

ServiceRole :: Maybe Text

jfdJobFlowId :: Lens' JobFlowDetail Text

The job flow identifier.

jfdName :: Lens' JobFlowDetail Text

The name of the job flow.

jfdLogUri :: Lens' JobFlowDetail (Maybe Text)

The location in Amazon S3 where log files for the job are stored.

jfdAmiVersion :: Lens' JobFlowDetail (Maybe Text)

The version of the AMI used to initialize Amazon EC2 instances in the job flow. For a list of AMI versions currently supported by Amazon ElasticMapReduce, go to AMI Versions Supported in Elastic MapReduce in the Amazon Elastic MapReduce Developer's Guide.

jfdExecutionStatusDetail :: Lens' JobFlowDetail JobFlowExecutionStatusDetail

Describes the execution status of the job flow.

jfdInstances :: Lens' JobFlowDetail JobFlowInstancesDetail

Describes the Amazon EC2 instances of the job flow.

jfdSteps :: Lens' JobFlowDetail [StepDetail]

A list of steps run by the job flow.

jfdBootstrapActions :: Lens' JobFlowDetail [BootstrapActionDetail]

A list of the bootstrap actions run by the job flow.

jfdSupportedProducts :: Lens' JobFlowDetail [Text]

A list of strings set by third party software when the job flow is launched. If you are not using third party software to manage the job flow this value is empty.

jfdVisibleToAllUsers :: Lens' JobFlowDetail (Maybe Bool)

Specifies whether the job flow is visible to all IAM users of the AWS account associated with the job flow. If this value is set to true, all IAM users of that AWS account can view and (if they have the proper policy permissions set) manage the job flow. If it is set to false, only the IAM user that created the job flow can view and manage it. This value can be changed using the SetVisibleToAllUsers action.

jfdJobFlowRole :: Lens' JobFlowDetail (Maybe Text)

The IAM role that was specified when the job flow was launched. The EC2 instances of the job flow assume this role.

jfdServiceRole :: Lens' JobFlowDetail (Maybe Text)

The IAM role that will be assumed by the Amazon EMR service to access AWS resources on your behalf.

JobFlowExecutionStatusDetail

data JobFlowExecutionStatusDetail

Describes the execution status of the job flow.

jobFlowExecutionStatusDetail

Arguments

:: JobFlowExecutionState
jfesdState

-> POSIX
jfesdCreationDateTime

-> JobFlowExecutionStatusDetail

Smart constructor for the minimum required fields to construct a valid JobFlowExecutionStatusDetail data type to populate a request.

The fields accessible through corresponding lenses are:

State :: JobFlowExecutionState

CreationDateTime :: POSIX

StartDateTime :: Maybe POSIX

ReadyDateTime :: Maybe POSIX

EndDateTime :: Maybe POSIX

LastStateChangeReason :: Maybe Text

jfesdState :: Lens' JobFlowExecutionStatusDetail JobFlowExecutionState

The state of the job flow.

jfesdCreationDateTime :: Lens' JobFlowExecutionStatusDetail POSIX

The creation date and time of the job flow.

jfesdStartDateTime :: Lens' JobFlowExecutionStatusDetail (Maybe POSIX)

The start date and time of the job flow.

jfesdReadyDateTime :: Lens' JobFlowExecutionStatusDetail (Maybe POSIX)

The date and time when the job flow was ready to start running bootstrap actions.

jfesdEndDateTime :: Lens' JobFlowExecutionStatusDetail (Maybe POSIX)

The completion date and time of the job flow.

jfesdLastStateChangeReason :: Lens' JobFlowExecutionStatusDetail (Maybe Text)

Description of the job flow last changed state.

JobFlowInstancesConfig

data JobFlowInstancesConfig

A specification of the number and type of Amazon EC2 instances on which to run the job flow.

jobFlowInstancesConfig :: JobFlowInstancesConfig

Smart constructor for the minimum required fields to construct a valid JobFlowInstancesConfig data type to populate a request.

The fields accessible through corresponding lenses are:

MasterInstanceType :: Maybe Text

SlaveInstanceType :: Maybe Text

InstanceCount :: Maybe Integer

InstanceGroups :: [InstanceGroupConfig]

Ec2KeyName :: Maybe Text

Placement :: Maybe PlacementType

KeepJobFlowAliveWhenNoSteps :: Maybe Bool

TerminationProtected :: Maybe Bool

HadoopVersion :: Maybe Text

Ec2SubnetId :: Maybe Text

jficMasterInstanceType :: Lens' JobFlowInstancesConfig (Maybe Text)

The EC2 instance type of the master node.

jficSlaveInstanceType :: Lens' JobFlowInstancesConfig (Maybe Text)

The EC2 instance type of the slave nodes.

jficInstanceCount :: Lens' JobFlowInstancesConfig (Maybe Integer)

The number of Amazon EC2 instances used to execute the job flow.

jficInstanceGroups :: Lens' JobFlowInstancesConfig [InstanceGroupConfig]

Configuration for the job flow's instance groups.

jficEc2KeyName :: Lens' JobFlowInstancesConfig (Maybe Text)

The name of the Amazon EC2 key pair that can be used to ssh to the master node as the user called "hadoop.".

jficPlacement :: Lens' JobFlowInstancesConfig (Maybe PlacementType)

The Availability Zone the job flow will run in.

jficKeepJobFlowAliveWhenNoSteps :: Lens' JobFlowInstancesConfig (Maybe Bool)

Specifies whether the job flow should terminate after completing all steps.

jficTerminationProtected :: Lens' JobFlowInstancesConfig (Maybe Bool)

Specifies whether to lock the job flow to prevent the Amazon EC2 instances from being terminated by API call, user intervention, or in the event of a job flow error.

jficHadoopVersion :: Lens' JobFlowInstancesConfig (Maybe Text)

The Hadoop version for the job flow. Valid inputs are "0.18", "0.20", or "0.20.205". If you do not set this value, the default of 0.18 is used, unless the AmiVersion parameter is set in the RunJobFlow call, in which case the default version of Hadoop for that AMI version is used.

jficEc2SubnetId :: Lens' JobFlowInstancesConfig (Maybe Text)

To launch the job flow in Amazon Virtual Private Cloud (Amazon VPC), set this parameter to the identifier of the Amazon VPC subnet where you want the job flow to launch. If you do not specify this value, the job flow is launched in the normal Amazon Web Services cloud, outside of an Amazon VPC. Amazon VPC currently does not support cluster compute quadruple extra large (cc1.4xlarge) instances. Thus you cannot specify the cc1.4xlarge instance type for nodes of a job flow launched in a Amazon VPC.

JobFlowInstancesDetail

data JobFlowInstancesDetail

Describes the Amazon EC2 instances of the job flow.

jobFlowInstancesDetail

Arguments

:: Text
jfidMasterInstanceType

-> Text
jfidSlaveInstanceType

-> Integer
jfidInstanceCount

-> JobFlowInstancesDetail

Smart constructor for the minimum required fields to construct a valid JobFlowInstancesDetail data type to populate a request.

The fields accessible through corresponding lenses are:

MasterInstanceType :: Text

MasterPublicDnsName :: Maybe Text

MasterInstanceId :: Maybe Text

SlaveInstanceType :: Text

InstanceCount :: Integer

InstanceGroups :: [InstanceGroupDetail]

NormalizedInstanceHours :: Maybe Integer

Ec2KeyName :: Maybe Text

Ec2SubnetId :: Maybe Text

Placement :: Maybe PlacementType

KeepJobFlowAliveWhenNoSteps :: Maybe Bool

TerminationProtected :: Maybe Bool

HadoopVersion :: Maybe Text

jfidMasterInstanceType :: Lens' JobFlowInstancesDetail Text

The Amazon EC2 master node instance type.

jfidMasterPublicDnsName :: Lens' JobFlowInstancesDetail (Maybe Text)

The DNS name of the master node.

jfidMasterInstanceId :: Lens' JobFlowInstancesDetail (Maybe Text)

The Amazon EC2 instance identifier of the master node.

jfidSlaveInstanceType :: Lens' JobFlowInstancesDetail Text

The Amazon EC2 slave node instance type.

jfidInstanceCount :: Lens' JobFlowInstancesDetail Integer

The number of Amazon EC2 instances in the cluster. If the value is 1, the same instance serves as both the master and slave node. If the value is greater than 1, one instance is the master node and all others are slave nodes.

jfidInstanceGroups :: Lens' JobFlowInstancesDetail [InstanceGroupDetail]

Details about the job flow's instance groups.

jfidNormalizedInstanceHours :: Lens' JobFlowInstancesDetail (Maybe Integer)

An approximation of the cost of the job flow, represented in m1.small/hours. This value is incremented once for every hour an m1.small runs. Larger instances are weighted more, so an Amazon EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

jfidEc2KeyName :: Lens' JobFlowInstancesDetail (Maybe Text)

The name of an Amazon EC2 key pair that can be used to ssh to the master node of job flow.

jfidEc2SubnetId :: Lens' JobFlowInstancesDetail (Maybe Text)

For job flows launched within Amazon Virtual Private Cloud, this value specifies the identifier of the subnet where the job flow was launched.

jfidPlacement :: Lens' JobFlowInstancesDetail (Maybe PlacementType)

The Amazon EC2 Availability Zone for the job flow.

jfidKeepJobFlowAliveWhenNoSteps :: Lens' JobFlowInstancesDetail (Maybe Bool)

Specifies whether the job flow should terminate after completing all steps.

jfidTerminationProtected :: Lens' JobFlowInstancesDetail (Maybe Bool)

Specifies whether the Amazon EC2 instances in the cluster are protected from termination by API calls, user intervention, or in the event of a job flow error.

jfidHadoopVersion :: Lens' JobFlowInstancesDetail (Maybe Text)

The Hadoop version for the job flow.

KeyValue

data KeyValue

A key value pair.

keyValue :: KeyValue

Smart constructor for the minimum required fields to construct a valid KeyValue data type to populate a request.

The fields accessible through corresponding lenses are:

Key :: Maybe Text

Value :: Maybe Text

kvKey :: Lens' KeyValue (Maybe Text)

The unique identifier of a key value pair.

kvValue :: Lens' KeyValue (Maybe Text)

The value part of the identified key.

ScriptBootstrapActionConfig

data ScriptBootstrapActionConfig

The script run by the bootstrap action.

scriptBootstrapActionConfig

Arguments

:: Text
sbacPath

-> ScriptBootstrapActionConfig

Smart constructor for the minimum required fields to construct a valid ScriptBootstrapActionConfig data type to populate a request.

The fields accessible through corresponding lenses are:

Path :: Text

Args :: [Text]

sbacPath :: Lens' ScriptBootstrapActionConfig Text

Location of the script to run during a bootstrap action. Can be either a location in Amazon S3 or on a local file system.

sbacArgs :: Lens' ScriptBootstrapActionConfig [Text]

A list of command line arguments to pass to the bootstrap action script.

Step

data Step

The step details for the requested step identifier.

step :: Step

Smart constructor for the minimum required fields to construct a valid Step data type.

Step is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Id :: Maybe Text

Name :: Maybe Text

Config :: Maybe HadoopStepConfig

ActionOnFailure :: Maybe ActionOnFailure

Status :: Maybe StepStatus

sId :: Lens' Step (Maybe Text)

The identifier of the cluster step.

sName :: Lens' Step (Maybe Text)

The name of the cluster step.

sConfig :: Lens' Step (Maybe HadoopStepConfig)

The Hadoop job configuration of the cluster step.

sActionOnFailure :: Lens' Step (Maybe ActionOnFailure)

This specifies what action to take when the cluster step fails. Possible values are TERMINATE_CLUSTER, CANCEL_AND_WAIT, and CONTINUE.

sStatus :: Lens' Step (Maybe StepStatus)

The current execution status details of the cluster step.

StepConfig

data StepConfig

Specification of a job flow step.

stepConfig

Arguments

:: Text
scName

-> HadoopJarStepConfig
scHadoopJarStep

-> StepConfig

Smart constructor for the minimum required fields to construct a valid StepConfig data type to populate a request.

The fields accessible through corresponding lenses are:

Name :: Text

ActionOnFailure :: Maybe ActionOnFailure

HadoopJarStep :: HadoopJarStepConfig

scName :: Lens' StepConfig Text

The name of the job flow step.

scActionOnFailure :: Lens' StepConfig (Maybe ActionOnFailure)

The action to take if the job flow step fails.

scHadoopJarStep :: Lens' StepConfig HadoopJarStepConfig

The JAR file used for the job flow step.

StepDetail

data StepDetail

Combines the execution state and configuration of a step.

stepDetail

Arguments

:: StepConfig
sdStepConfig

-> StepExecutionStatusDetail
sdExecutionStatusDetail

-> StepDetail

Smart constructor for the minimum required fields to construct a valid StepDetail data type to populate a request.

The fields accessible through corresponding lenses are:

StepConfig :: StepConfig

ExecutionStatusDetail :: StepExecutionStatusDetail

sdStepConfig :: Lens' StepDetail StepConfig

The step configuration.

sdExecutionStatusDetail :: Lens' StepDetail StepExecutionStatusDetail

The description of the step status.

StepExecutionStatusDetail

data StepExecutionStatusDetail

The description of the step status.

stepExecutionStatusDetail

Arguments

:: StepExecutionState
sesdState

-> POSIX
sesdCreationDateTime

-> StepExecutionStatusDetail

Smart constructor for the minimum required fields to construct a valid StepExecutionStatusDetail data type to populate a request.

The fields accessible through corresponding lenses are:

State :: StepExecutionState

CreationDateTime :: POSIX

StartDateTime :: Maybe POSIX

EndDateTime :: Maybe POSIX

LastStateChangeReason :: Maybe Text

sesdState :: Lens' StepExecutionStatusDetail StepExecutionState

The state of the job flow step.

sesdCreationDateTime :: Lens' StepExecutionStatusDetail POSIX

The creation date and time of the step.

sesdStartDateTime :: Lens' StepExecutionStatusDetail (Maybe POSIX)

The start date and time of the step.

sesdEndDateTime :: Lens' StepExecutionStatusDetail (Maybe POSIX)

The completion date and time of the step.

sesdLastStateChangeReason :: Lens' StepExecutionStatusDetail (Maybe Text)

A description of the step's current state.

StepStateChangeReason

data StepStateChangeReason

The reason for the step execution status change.

stepStateChangeReason :: StepStateChangeReason

Smart constructor for the minimum required fields to construct a valid StepStateChangeReason data type to populate a request.

The fields accessible through corresponding lenses are:

Code :: Maybe StepStateChangeReasonCode

Message :: Maybe Text

sscrCode :: Lens' StepStateChangeReason (Maybe StepStateChangeReasonCode)

The programmable code for the state change reason.

sscrMessage :: Lens' StepStateChangeReason (Maybe Text)

The descriptive message for the state change reason.

StepStatus

data StepStatus

The current execution status details of the cluster step.

stepStatus :: StepStatus

Smart constructor for the minimum required fields to construct a valid StepStatus data type to populate a request.

The fields accessible through corresponding lenses are:

State :: Maybe StepState

StateChangeReason :: Maybe StepStateChangeReason

Timeline :: Maybe StepTimeline

ssState :: Lens' StepStatus (Maybe StepState)

The execution state of the cluster step.

ssStateChangeReason :: Lens' StepStatus (Maybe StepStateChangeReason)

The reason for the step execution status change.

ssTimeline :: Lens' StepStatus (Maybe StepTimeline)

The timeline of the cluster step status over time.

StepSummary

data StepSummary

The summary of the cluster step.

stepSummary :: StepSummary

Smart constructor for the minimum required fields to construct a valid StepSummary data type.

StepSummary is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Id :: Maybe Text

Name :: Maybe Text

Status :: Maybe StepStatus

ssrId :: Lens' StepSummary (Maybe Text)

The identifier of the cluster step.

ssrName :: Lens' StepSummary (Maybe Text)

The name of the cluster step.

ssrStatus :: Lens' StepSummary (Maybe StepStatus)

The current execution status details of the cluster step.

StepTimeline

data StepTimeline

The timeline of the cluster step status over time.

stepTimeline :: StepTimeline

Smart constructor for the minimum required fields to construct a valid StepTimeline data type to populate a request.

The fields accessible through corresponding lenses are:

CreationDateTime :: Maybe POSIX

StartDateTime :: Maybe POSIX

EndDateTime :: Maybe POSIX

stCreationDateTime :: Lens' StepTimeline (Maybe POSIX)

The date and time when the cluster step was created.

stStartDateTime :: Lens' StepTimeline (Maybe POSIX)

The date and time when the cluster step execution started.

stEndDateTime :: Lens' StepTimeline (Maybe POSIX)

The date and time when the cluster step execution completed or failed.

SupportedProductConfig

data SupportedProductConfig

The list of supported product configurations which allow user-supplied arguments. EMR accepts these arguments and forwards them to the corresponding installation script as bootstrap action arguments.

supportedProductConfig :: SupportedProductConfig

Smart constructor for the minimum required fields to construct a valid SupportedProductConfig data type to populate a request.

The fields accessible through corresponding lenses are:

Name :: Maybe Text

Args :: [Text]

spcName :: Lens' SupportedProductConfig (Maybe Text)

The name of the product configuration.

spcArgs :: Lens' SupportedProductConfig [Text]

The list of user-supplied arguments.

Tag

data Tag

A key/value pair containing user-defined metadata that you can associate with an Amazon EMR resource. Tags make it easier to associate clusters in various ways, such as grouping clu sters to track your Amazon EMR resource allocation costs. For more information, see Tagging Amazon EMR Resources.

tag :: Tag

Smart constructor for the minimum required fields to construct a valid Tag data type to populate a request.

The fields accessible through corresponding lenses are:

Key :: Maybe Text

Value :: Maybe Text

tKey :: Lens' Tag (Maybe Text)

A user-defined key, which is the minimum required information for a valid tag. For more information, see Tagging Amazon EMR Resources.

tValue :: Lens' Tag (Maybe Text)

A user-defined value, which is optional in a tag. For more information, see Tagging Amazon EMR Resources.