Network.AWS.EMR.V2009_03_31.DescribeCluster

Description

Provides cluster-level details including status, hardware and software configuration, VPC settings, and so on. For information about the cluster steps, see ListSteps.

Documentation

data DescribeCluster

Constructors

DescribeCluster

Fields

_dciClusterId :: Text

The identifier of the cluster to describe.

dciClusterId :: Iso' DescribeCluster Text

data DescribeClusterResponse

Constructors

DescribeClusterResponse

Fields

_dcoCluster :: Maybe Cluster

This output contains the details for the requested cluster.

dcoCluster :: Iso' DescribeClusterResponse (Maybe Cluster)