Network.AWS.EMR.V2009_03_31.DescribeJobFlows

Description

This API is deprecated and will eventually be removed. We recommend you use ListClusters, DescribeCluster, ListSteps, ListInstanceGroups and ListBootstrapActions instead. DescribeJobFlows returns a list of job flows that match all of the supplied parameters. The parameters can include a list of job flow IDs, job flow states, and restrictions on job flow creation date and time. Regardless of supplied parameters, only job flows created within the last two months are returned. If no parameters are supplied, then job flows matching either of the following criteria are returned: Job flows created and completed in the last two weeks Job flows created within the last two months that are in one of the following states: RUNNING, WAITING, SHUTTING_DOWN, STARTING Amazon Elastic MapReduce can return a maximum of 512 job flow descriptions. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: ElasticMapReduce.DescribeJobFlows Content-Length: 62 User-Agent: aws-sdk-ruby1.9.2 ruby1.9.3 i386-mingw32 Host: us-east-1.elasticmapreduce.amazonaws.com X-Amz-Date: 20130715T220330Z X-Amz-Content-Sha256: fce83af973f96f173512aca2845c56862b946feb1de0600326f1365b658a0e39 Authorization: AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20130715us-east-1elasticmapreduceaws4_request,

SignedHeaders=content-length;content-type;host;user-agent;x-amz-content-sha256;x-amz-date;x-amz-target, Signature=29F98a6f44e05ad54fe1e8b3d1a7101ab08dc3ad348995f89c533693cee2bb3b Accept: */* { JobFlowIds: ["j-ZKIY4CKQRX72"], DescriptionType: EXTENDED } HTTP/1.1 200 OK x-amzn-RequestId: 634d4142-ed9a-11e2-bbba-b56d7d016ec4 Content-Type: application/x-amz-json-1.1 Content-Length: 1624 Date: Mon, 15 Jul 2013 22:03:31 GMT {JobFlows: [{ AmiVersion: "2.3.6", BootstrapActions: [], ExecutionStatusDetail: { CreationDateTime: 1.373923429E9, EndDateTime: 1.373923995E9, LastStateChangeReason: "Steps completed", ReadyDateTime: 1.373923754E9, StartDateTime: 1.373923754E9, State: COMPLETED }, Instances: { HadoopVersion: "1.0.3", InstanceCount: 1, InstanceGroups: [{ CreationDateTime: 1.373923429E9, EndDateTime: 1.373923995E9, InstanceGroupId: "ig-3SRUWV3E0NB7K", InstanceRequestCount: 1, InstanceRole: MASTER, InstanceRunningCount: 0, InstanceType: "m1.small", LastStateChangeReason: "Job flow terminated", Market: ON_DEMAND, Name: "Master InstanceGroup", ReadyDateTime: 1.37392375E9, StartDateTime: 1.373923646E9, State: ENDED }], KeepJobFlowAliveWhenNoSteps: false, MasterInstanceId: "i-8c4fbbef", MasterInstanceType: "m1.small", MasterPublicDnsName: "ec2-107-20-46-140.compute-1.amazonaws.com", NormalizedInstanceHours: 1, Placement: {AvailabilityZone: "us-east-1a"}, TerminationProtected: false }, JobFlowId: "j-ZKIY4CKQRX72", Name: "Development Job Flow", Steps: [{ ExecutionStatusDetail: { CreationDateTime: 1.373923429E9, EndDateTime: 1.373923914E9, StartDateTime: 1.373923754E9, State: COMPLETED }, StepConfig: { ActionOnFailure: CANCEL_AND_WAIT, HadoopJarStep: { Args: [ "-input", "s3:/elasticmapreducesampleswordcountinput", "-output", "s3:/examples-bucketexample-output", "-mapper", "s3:/elasticmapreducesampleswordcountwordSplitter.py", "-reducer", "aggregate" ], Jar: "homehadoopcontribstreaming/hadoop-streaming.jar", Properties: [] }, Name: "Example Streaming Step" } }], SupportedProducts: [], VisibleToAllUsers: false }]}.

Request

data DescribeJobFlows

The input for the DescribeJobFlows operation.

Request constructor

mkDescribeJobFlows :: DescribeJobFlows

Smart constructor for the minimum required parameters to construct a valid DescribeJobFlows request.

Request lenses

djfCreatedAfter :: Lens' DescribeJobFlows (Maybe POSIX)

Return only job flows created after this date and time.

djfCreatedBefore :: Lens' DescribeJobFlows (Maybe POSIX)

Return only job flows created before this date and time.

djfJobFlowIds :: Lens' DescribeJobFlows [Text]

Return only job flows whose job flow ID is contained in this list.

djfJobFlowStates :: Lens' DescribeJobFlows [JobFlowExecutionState]

Return only job flows whose state is contained in this list.

Response

data DescribeJobFlowsResponse

The output for the DescribeJobFlows operation.

Response constructor

mkDescribeJobFlowsResponse :: DescribeJobFlowsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeJobFlowsResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

djfrJobFlows :: Lens' DescribeJobFlowsResponse [JobFlowDetail]

A list of job flows matching the parameters supplied.