Network.AWS.EMR.V2009_03_31.ListBootstrapActions

Description

Provides information about the bootstrap actions associated with a cluster.

Request

data ListBootstrapActions

This input determines which bootstrap actions to retrieve.

Request constructor

mkListBootstrapActions

Arguments

:: Text
lbaClusterId

-> ListBootstrapActions

Smart constructor for the minimum required parameters to construct a valid ListBootstrapActions request.

Request lenses

lbaClusterId :: Lens' ListBootstrapActions Text

The cluster identifier for the bootstrap actions to list .

lbaMarker :: Lens' ListBootstrapActions (Maybe Text)

The pagination token that indicates the next set of results to retrieve .

Response

data ListBootstrapActionsResponse

This output contains the boostrap actions detail .

Response constructor

mkListBootstrapActionsResponse :: ListBootstrapActionsResponse

Smart constructor for the minimum required parameters to construct a valid ListBootstrapActionsResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

lbarBootstrapActions :: Lens' ListBootstrapActionsResponse [Command]

The bootstrap actions associated with the cluster .

lbarMarker :: Lens' ListBootstrapActionsResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve .