Network.AWS.EMR.V2009_03_31.ListInstances

Description

Provides information about the cluster instances that Amazon EMR provisions on behalf of a user when it creates the cluster. For example, this operation indicates when the EC2 instances reach the Ready state, when instances become available to Amazon EMR to use for jobs, and the IP addresses for cluster instances, etc.

Request

data ListInstances

This input determines which instances to list.

Request constructor

mkListInstances

Arguments

:: Text
liClusterId

-> ListInstances

Smart constructor for the minimum required parameters to construct a valid ListInstances request.

The fields accessible through corresponding lenses are:

ClusterId :: Text

InstanceGroupId :: Maybe Text

InstanceGroupTypes :: [InstanceGroupType]

Marker :: Maybe Text

Request lenses

liClusterId :: Lens' ListInstances Text

The identifier of the cluster for which to list the instances.

liInstanceGroupId :: Lens' ListInstances (Maybe Text)

The identifier of the instance group for which to list the instances.

liInstanceGroupTypes :: Lens' ListInstances [InstanceGroupType]

The type of instance group for which to list the instances.

liMarker :: Lens' ListInstances (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

Response

data ListInstancesResponse

This output contains the list of instances.

Response constructor

mkListInstancesResponse :: ListInstancesResponse

Smart constructor for the minimum required parameters to construct a valid ListInstancesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Instances :: [Instance]

Marker :: Maybe Text

Response lenses

lirInstances :: Lens' ListInstancesResponse [Instance]

The list of instances for the cluster and given filters.

lirMarker :: Lens' ListInstancesResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.