Network.AWS.EMR.V2009_03_31.ModifyInstanceGroups

Description

ModifyInstanceGroups modifies the number of nodes and configuration settings of an instance group. The input parameters include the new target instance count for the group and the instance group ID. The call will either succeed or fail atomically. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: ElasticMapReduce.ModifyInstanceGroups Content-Length: 77 User-Agent: aws-sdk-ruby1.9.2 ruby1.9.3 i386-mingw32 Host: us-east-1.elasticmapreduce.amazonaws.com X-Amz-Date: 20130716T205843Z X-Amz-Content-Sha256: bb1af3d0c6c6a1a09f21ccd7f04a0e2e6c9ce5b5810b0f6777560fe4f81bda8c Authorization: AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20130716us-east-1elasticmapreduceaws4_request,

SignedHeaders=content-length;content-type;host;user-agent;x-amz-content-sha256;x-amz-date;x-amz-target, Signature=17bbbb4448a1f47a14d5657445e9de5cadf16bed58b850585f80865882133b33 Accept: */* {InstanceGroups: [{ InstanceGroupId: "ig-1S8NWT31S2OVG", InstanceCount: 5 }]} HTTP/1.1 200 OK x-amzn-RequestId: 80a74808-ee5a-11e2-90db-69a5154aeb8d Content-Type: application/x-amz-json-1.1 Content-Length: 0 Date: Tue, 16 Jul 2013 20:58:44 GMT.

Request

data ModifyInstanceGroups

Change the size of some instance groups.

Request constructor

mkModifyInstanceGroups :: ModifyInstanceGroups

Smart constructor for the minimum required parameters to construct a valid ModifyInstanceGroups request.

Request lenses

migInstanceGroups :: Lens' ModifyInstanceGroups [InstanceGroupModifyConfig]

Instance groups to change.

Response

data ModifyInstanceGroupsResponse