Network.AWS.EMR.V2009_03_31.RemoveTags

Description

Removes tags from an Amazon EMR resource. Tags make it easier to associate clusters in various ways, such as grouping clusters to track your Amazon EMR resource allocation costs. For more information, see Tagging Amazon EMR Resources. The following example removes the stack tag with value Prod from a cluster: POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: ElasticMapReduce.RemoveTags AUTHPARAMS { ResourceId: "j-3U7TSX5GZFD8Y", Tags: [{ Key: "stack", Value: Prod }] } HTTP/1.1 200 OK x-amzn-RequestId: 9da5a349-ed9e-11e2-90db-69a5154aeb8d Content-Type: application/x-amz-json-1.1 Content-Length: 71 Date: Mon, 15 Jul 2013 22:33:47 GMT { } The following example removes the stack and hbase tags from a cluster: POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: ElasticMapReduce.RemoveTags AUTHPARAMS { ResourceId: "j-3U7TSX5GZFD8Y", Tags: [{ Key: "stack" }, { Key: "hbase" }] } HTTP/1.1 200 OK x-amzn-RequestId: 9da5a349-ed9e-11e2-90db-69a5154aeb8d Content-Type: application/x-amz-json-1.1 Content-Length: 71 Date: Mon, 15 Jul 2013 22:33:47 GMT { }.

Request

data RemoveTags

This input identifies a cluster and a list of tags to remove.

Request constructor

mkRemoveTags

Arguments

:: Text
rtResourceId

-> [Text]
rtTagKeys

-> RemoveTags

Smart constructor for the minimum required parameters to construct a valid RemoveTags request.

Request lenses

rtResourceId :: Lens' RemoveTags Text

The Amazon EMR resource identifier from which tags will be removed. This value must be a cluster identifier.

rtTagKeys :: Lens' RemoveTags [Text]

A list of tag keys to remove from a resource.

Response

data RemoveTagsResponse

This output indicates the result of removing tags from a resource.