Network.AWS.EMR.V2009_03_31.SetTerminationProtection

Description

SetTerminationProtection locks a job flow so the Amazon EC2 instances in the cluster cannot be terminated by user intervention, an API call, or in the event of a job-flow error. The cluster still terminates upon successful completion of the job flow. Calling SetTerminationProtection on a job flow is analogous to calling the Amazon EC2 DisableAPITermination API on all of the EC2 instances in a cluster. SetTerminationProtection is used to prevent accidental termination of a job flow and to ensure that in the event of an error, the instances will persist so you can recover any data stored in their ephemeral instance storage. To terminate a job flow that has been locked by setting SetTerminationProtection to true, you must first unlock the job flow by a subsequent call to SetTerminationProtection in which you set the value to false. For more information, go to Protecting a Job Flow from Termination in the Amazon Elastic MapReduce Developer's Guide. POST / HTTP1.1 Content-Type: applicationx-amz-json-1.1 X-Amz-Target: ElasticMapReduce.SetTerminationProtection Content-Length: 61 User-Agent: aws-sdk-ruby1.9.2 ruby1.9.3 i386-mingw32 Host: us-east-1.elasticmapreduce.amazonaws.com X-Amz-Date: 20130716T211420Z X-Amz-Content-Sha256: c362fadae0fce377aa63f04388aeb90c53cedb17a8bfbb8cffcb10c2378137f9 Authorization: AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20130716us-east-1elasticmapreduceaws4_request,

SignedHeaders=content-length;content-type;host;user-agent;x-amz-content-sha256;x-amz-date;x-amz-target, Signature=764b6aa1a38733cadff35a2e884887e9f1208a422266bc83ac77e8d0b80bd4cf Accept: */* { JobFlowIds: ["j-3TS0OIYO4NFN"], TerminationProtected: true } HTTP/1.1 200 OK x-amzn-RequestId: af23b1db-ee5c-11e2-9787-192218ecb460 Content-Type: application/x-amz-json-1.1 Content-Length: 0 Date: Tue, 16 Jul 2013 21:14:21 GMT.

Request

data SetTerminationProtection

The input argument to the TerminationProtection operation.

Request constructor

mkSetTerminationProtection

Arguments

:: [Text]
stpJobFlowIds

-> Bool
stpTerminationProtected

-> SetTerminationProtection

Smart constructor for the minimum required parameters to construct a valid SetTerminationProtection request.

The fields accessible through corresponding lenses are:

JobFlowIds :: [Text]

TerminationProtected :: Bool

Request lenses

stpJobFlowIds :: Lens' SetTerminationProtection [Text]

A list of strings that uniquely identify the job flows to protect. This identifier is returned by RunJobFlow and can also be obtained from DescribeJobFlows .

stpTerminationProtected :: Lens' SetTerminationProtection Bool

A Boolean that indicates whether to protect the job flow and prevent the Amazon EC2 instances in the cluster from shutting down due to API calls, user intervention, or job-flow error.

Response

data SetTerminationProtectionResponse

Response constructor

mkSetTerminationProtectionResponse :: SetTerminationProtectionResponse

Smart constructor for the minimum required parameters to construct a valid SetTerminationProtectionResponse response.

This constructor is provided for convenience and testing purposes.