Network.AWS.EMR.AddInstanceGroups

Description

Adds one or more instance groups to a running cluster.

Creating a Request

addInstanceGroups

Arguments

:: Text
aigJobFlowId

-> AddInstanceGroups

Creates a value of AddInstanceGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aigInstanceGroups - Instance groups to add.

aigJobFlowId - Job flow in which to add the instance groups.

data AddInstanceGroups

Input to an AddInstanceGroups call.

See: addInstanceGroups smart constructor.

Request Lenses

aigInstanceGroups :: Lens' AddInstanceGroups [InstanceGroupConfig]

Instance groups to add.

aigJobFlowId :: Lens' AddInstanceGroups Text

Job flow in which to add the instance groups.

Destructuring the Response

addInstanceGroupsResponse

Arguments

:: Int
aigrsResponseStatus

-> AddInstanceGroupsResponse

Creates a value of AddInstanceGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aigrsJobFlowId - The job flow ID in which the instance groups are added.

aigrsInstanceGroupIds - Instance group IDs of the newly created instance groups.

aigrsResponseStatus - -- | The response status code.

data AddInstanceGroupsResponse

Output from an AddInstanceGroups call.

See: addInstanceGroupsResponse smart constructor.

Response Lenses

aigrsJobFlowId :: Lens' AddInstanceGroupsResponse (Maybe Text)

The job flow ID in which the instance groups are added.

aigrsInstanceGroupIds :: Lens' AddInstanceGroupsResponse [Text]

Instance group IDs of the newly created instance groups.

aigrsResponseStatus :: Lens' AddInstanceGroupsResponse Int

- | The response status code.