Network.AWS.EMR.AddJobFlowSteps

Description

AddJobFlowSteps adds new steps to a running cluster. A maximum of 256 steps are allowed in each job flow.

If your cluster is long-running (such as a Hive data warehouse) or complex, you may require more than 256 steps to process your data. You can bypass the 256-step limitation in various ways, including using SSH to connect to the master node and submitting queries directly to the software running on the master node, such as Hive and Hadoop. For more information on how to do this, see Add More than 256 Steps to a Cluster in the Amazon EMR Management Guide .

A step specifies the location of a JAR file stored either on the master node of the cluster or in Amazon S3. Each step is performed by the main function of the main class of the JAR file. The main class can be specified either in the manifest of the JAR or by using the MainFunction parameter of the step.

Amazon EMR executes each step in the order listed. For a step to be considered complete, the main function must exit with a zero exit code and all Hadoop jobs started while the step was running must have completed and run successfully.

You can only add steps to a cluster that is in one of the following states: STARTING, BOOTSTRAPPING, RUNNING, or WAITING.

Creating a Request

addJobFlowSteps

Arguments

:: Text
ajfsJobFlowId

-> AddJobFlowSteps

Creates a value of AddJobFlowSteps with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ajfsJobFlowId - A string that uniquely identifies the job flow. This identifier is returned by RunJobFlow and can also be obtained from ListClusters .

ajfsSteps - A list of StepConfig to be executed by the job flow.

data AddJobFlowSteps

The input argument to the AddJobFlowSteps operation.

See: addJobFlowSteps smart constructor.

Request Lenses

ajfsJobFlowId :: Lens' AddJobFlowSteps Text

A string that uniquely identifies the job flow. This identifier is returned by RunJobFlow and can also be obtained from ListClusters .

ajfsSteps :: Lens' AddJobFlowSteps [StepConfig]

A list of StepConfig to be executed by the job flow.

Destructuring the Response

addJobFlowStepsResponse

Arguments

:: Int
ajfsrsResponseStatus

-> AddJobFlowStepsResponse

Creates a value of AddJobFlowStepsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ajfsrsStepIds - The identifiers of the list of steps added to the job flow.

ajfsrsResponseStatus - -- | The response status code.

data AddJobFlowStepsResponse

The output for the AddJobFlowSteps operation.

See: addJobFlowStepsResponse smart constructor.

Response Lenses

ajfsrsStepIds :: Lens' AddJobFlowStepsResponse [Text]

The identifiers of the list of steps added to the job flow.

ajfsrsResponseStatus :: Lens' AddJobFlowStepsResponse Int

- | The response status code.