Network.AWS.EMR.ListClusters

Description

Provides the status of all clusters visible to this AWS account. Allows you to filter the list of clusters based on certain criteria; for example, filtering by cluster creation date and time or by status. This call returns a maximum of 50 clusters per call, but returns a marker to track the paging of the cluster list across multiple ListClusters calls.

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_ListClusters.html

Request

data ListClusters

Request constructor

listClusters :: ListClusters

ListClusters constructor.

The fields accessible through corresponding lenses are:

lcClusterStates :: [ClusterState]

lcCreatedAfter :: Maybe UTCTime

lcCreatedBefore :: Maybe UTCTime

lcMarker :: Maybe Text

Request lenses

lcClusterStates :: Lens' ListClusters [ClusterState]

The cluster state filters to apply when listing clusters.

lcCreatedAfter :: Lens' ListClusters (Maybe UTCTime)

The creation date and time beginning value filter for listing clusters .

lcCreatedBefore :: Lens' ListClusters (Maybe UTCTime)

The creation date and time end value filter for listing clusters .

lcMarker :: Lens' ListClusters (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

Response

data ListClustersResponse

Response constructor

listClustersResponse :: ListClustersResponse

ListClustersResponse constructor.

The fields accessible through corresponding lenses are:

lcrClusters :: [ClusterSummary]

lcrMarker :: Maybe Text

Response lenses

lcrClusters :: Lens' ListClustersResponse [ClusterSummary]

The list of clusters for the account based on the given filters.

lcrMarker :: Lens' ListClustersResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.