Network.AWS.EMR.ListInstanceGroups

Description

Provides all available details about the instance groups in a cluster.

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_ListInstanceGroups.html

Request

data ListInstanceGroups

Request constructor

listInstanceGroups

Arguments

:: Text
ligClusterId

-> ListInstanceGroups

ListInstanceGroups constructor.

The fields accessible through corresponding lenses are:

ligClusterId :: Text

ligMarker :: Maybe Text

Request lenses

ligClusterId :: Lens' ListInstanceGroups Text

The identifier of the cluster for which to list the instance groups.

ligMarker :: Lens' ListInstanceGroups (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

Response

data ListInstanceGroupsResponse

Response constructor

listInstanceGroupsResponse :: ListInstanceGroupsResponse

ListInstanceGroupsResponse constructor.

The fields accessible through corresponding lenses are:

ligrInstanceGroups :: [InstanceGroup]

ligrMarker :: Maybe Text

Response lenses

ligrInstanceGroups :: Lens' ListInstanceGroupsResponse [InstanceGroup]

The list of instance groups for the cluster and given filters.

ligrMarker :: Lens' ListInstanceGroupsResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.