Network.AWS.EMR.ListSteps

Description

Provides a list of steps for the cluster in reverse order unless you specify stepIds with the request.

This operation returns paginated results.

Creating a Request

listSteps

Arguments

:: Text
lsClusterId

-> ListSteps

Creates a value of ListSteps with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsStepIds - The filter to limit the step list based on the identifier of the steps.

lsStepStates - The filter to limit the step list based on certain states.

lsMarker - The pagination token that indicates the next set of results to retrieve.

lsClusterId - The identifier of the cluster for which to list the steps.

data ListSteps

This input determines which steps to list.

See: listSteps smart constructor.

Request Lenses

lsStepIds :: Lens' ListSteps [Text]

The filter to limit the step list based on the identifier of the steps.

lsStepStates :: Lens' ListSteps [StepState]

The filter to limit the step list based on certain states.

lsMarker :: Lens' ListSteps (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

lsClusterId :: Lens' ListSteps Text

The identifier of the cluster for which to list the steps.

Destructuring the Response

listStepsResponse

Arguments

:: Int
lsrsResponseStatus

-> ListStepsResponse

Creates a value of ListStepsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsrsSteps - The filtered list of steps for the cluster.

lsrsMarker - The pagination token that indicates the next set of results to retrieve.

lsrsResponseStatus - -- | The response status code.

data ListStepsResponse

This output contains the list of steps returned in reverse order. This means that the last step is the first element in the list.

See: listStepsResponse smart constructor.

Response Lenses

lsrsSteps :: Lens' ListStepsResponse [StepSummary]

The filtered list of steps for the cluster.

lsrsMarker :: Lens' ListStepsResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

lsrsResponseStatus :: Lens' ListStepsResponse Int

- | The response status code.