Network.AWS.EMR.RunJobFlow

Description

RunJobFlow creates and starts running a new cluster (job flow). The cluster runs the steps specified. After the steps complete, the cluster stops and the HDFS partition is lost. To prevent loss of data, configure the last step of the job flow to store results in Amazon S3. If the JobFlowInstancesConfig KeepJobFlowAliveWhenNoSteps parameter is set to TRUE , the cluster transitions to the WAITING state rather than shutting down after the steps have completed.

For additional protection, you can set the JobFlowInstancesConfig TerminationProtected parameter to TRUE to lock the cluster and prevent it from being terminated by API call, user intervention, or in the event of a job flow error.

A maximum of 256 steps are allowed in each job flow.

If your cluster is long-running (such as a Hive data warehouse) or complex, you may require more than 256 steps to process your data. You can bypass the 256-step limitation in various ways, including using the SSH shell to connect to the master node and submitting queries directly to the software running on the master node, such as Hive and Hadoop. For more information on how to do this, see Add More than 256 Steps to a Cluster in the Amazon EMR Management Guide .

For long running clusters, we recommend that you periodically store your results.

Creating a Request

runJobFlow

Arguments

:: Text
rjfName

-> JobFlowInstancesConfig
rjfInstances

-> RunJobFlow

Creates a value of RunJobFlow with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rjfAMIVersion - For Amazon EMR AMI versions 3.x and 2.x. For Amazon EMR releases 4.0 and later, the Linux AMI is determined by the ReleaseLabel specified or by CustomAmiID . The version of the Amazon Machine Image (AMI) to use when launching Amazon EC2 instances in the job flow. For details about the AMI versions currently supported in EMR version 3.x and 2.x, see AMI Versions Supported in EMR in the Amazon EMR Developer Guide . If the AMI supports multiple versions of Hadoop (for example, AMI 1.0 supports both Hadoop 0.18 and 0.20), you can use the JobFlowInstancesConfig HadoopVersion parameter to modify the version of Hadoop from the defaults shown above.

rjfEBSRootVolumeSize - The size, in GiB, of the EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

rjfAdditionalInfo - A JSON string for selecting additional features.

rjfConfigurations - For Amazon EMR releases 4.0 and later. The list of configurations supplied for the EMR cluster you are creating.

rjfCustomAMIId - Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI. If specified, Amazon EMR uses this AMI when it launches cluster EC2 instances. For more information about custom AMIs in Amazon EMR, see Using a Custom AMI in the Amazon EMR Management Guide . If omitted, the cluster uses the base Linux AMI for the ReleaseLabel specified. For Amazon EMR versions 2.x and 3.x, use AmiVersion instead. For information about creating a custom AMI, see Creating an Amazon EBS-Backed Linux AMI in the Amazon Elastic Compute Cloud User Guide for Linux Instances . For information about finding an AMI ID, see Finding a Linux AMI .

rjfAutoScalingRole - An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole . The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

rjfSecurityConfiguration - The name of a security configuration to apply to the cluster.

rjfScaleDownBehavior - Specifies the way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR blacklists and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

rjfSteps - A list of steps to run.

rjfJobFlowRole - Also called instance profile and EC2 role. An IAM role for an EMR cluster. The EC2 instances of the cluster assume this role. The default role is EMR_EC2_DefaultRole . In order to use the default role, you must have already created it using the CLI or console.

rjfBootstrapActions - A list of bootstrap actions to run before Hadoop starts on the cluster nodes.

rjfReleaseLabel - The release label for the Amazon EMR release. For Amazon EMR 3.x and 2.x AMIs, use AmiVersion instead.

rjfRepoUpgradeOnBoot - Applies only when CustomAmiID is used. Specifies which updates from the Amazon Linux AMI package repositories to apply automatically when the instance boots using the AMI. If omitted, the default is SECURITY , which indicates that only security updates are applied. If NONE is specified, no updates are applied, and all updates must be applied manually.

rjfLogURI - The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

rjfNewSupportedProducts - A list of strings that indicates third-party software to use with the job flow that accepts a user argument list. EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action arguments. For more information, see "Launch a Job Flow on the MapR Distribution for Hadoop" in the Amazon EMR Developer Guide . Supported values are: * "mapr-m3" - launch the cluster using MapR M3 Edition. * "mapr-m5" - launch the cluster using MapR M5 Edition. * "mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition respectively. * "mapr-m7" - launch the cluster using MapR M7 Edition. * "hunk" - launch the cluster with the Hunk Big Data Analtics Platform. * "hue"- launch the cluster with Hue installed. * "spark" - launch the cluster with Apache Spark installed. * "ganglia" - launch the cluster with the Ganglia Monitoring System installed.

rjfVisibleToAllUsers - Whether the cluster is visible to all IAM users of the AWS account associated with the cluster. If this value is set to true , all IAM users of that AWS account can view and (if they have the proper policy permissions set) manage the cluster. If it is set to false , only the IAM user that created the cluster can view and manage it.

rjfSupportedProducts - A list of strings that indicates third-party software to use. For more information, see Use Third Party Applications with Amazon EMR . Currently supported values are: * "mapr-m3" - launch the job flow using MapR M3 Edition. * "mapr-m5" - launch the job flow using MapR M5 Edition.

rjfApplications - For Amazon EMR releases 4.0 and later. A list of applications for the cluster. Valid values are: Hadoop, Hive, Mahout, Pig, and "Spark." They are case insensitive.

rjfTags - A list of tags to associate with a cluster and propagate to Amazon EC2 instances.

rjfServiceRole - The IAM role that will be assumed by the Amazon EMR service to access AWS resources on your behalf.

rjfName - The name of the job flow.

rjfInstances - A specification of the number and type of Amazon EC2 instances.

data RunJobFlow

Input to the RunJobFlow operation.

See: runJobFlow smart constructor.

Request Lenses

rjfAMIVersion :: Lens' RunJobFlow (Maybe Text)

For Amazon EMR AMI versions 3.x and 2.x. For Amazon EMR releases 4.0 and later, the Linux AMI is determined by the ReleaseLabel specified or by CustomAmiID . The version of the Amazon Machine Image (AMI) to use when launching Amazon EC2 instances in the job flow. For details about the AMI versions currently supported in EMR version 3.x and 2.x, see AMI Versions Supported in EMR in the Amazon EMR Developer Guide . If the AMI supports multiple versions of Hadoop (for example, AMI 1.0 supports both Hadoop 0.18 and 0.20), you can use the JobFlowInstancesConfig HadoopVersion parameter to modify the version of Hadoop from the defaults shown above.

rjfEBSRootVolumeSize :: Lens' RunJobFlow (Maybe Int)

The size, in GiB, of the EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

rjfAdditionalInfo :: Lens' RunJobFlow (Maybe Text)

A JSON string for selecting additional features.

rjfConfigurations :: Lens' RunJobFlow [Configuration]

For Amazon EMR releases 4.0 and later. The list of configurations supplied for the EMR cluster you are creating.

rjfCustomAMIId :: Lens' RunJobFlow (Maybe Text)

Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI. If specified, Amazon EMR uses this AMI when it launches cluster EC2 instances. For more information about custom AMIs in Amazon EMR, see Using a Custom AMI in the Amazon EMR Management Guide . If omitted, the cluster uses the base Linux AMI for the ReleaseLabel specified. For Amazon EMR versions 2.x and 3.x, use AmiVersion instead. For information about creating a custom AMI, see Creating an Amazon EBS-Backed Linux AMI in the Amazon Elastic Compute Cloud User Guide for Linux Instances . For information about finding an AMI ID, see Finding a Linux AMI .

rjfAutoScalingRole :: Lens' RunJobFlow (Maybe Text)

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole . The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

rjfSecurityConfiguration :: Lens' RunJobFlow (Maybe Text)

The name of a security configuration to apply to the cluster.

rjfScaleDownBehavior :: Lens' RunJobFlow (Maybe ScaleDownBehavior)

Specifies the way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR blacklists and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

rjfSteps :: Lens' RunJobFlow [StepConfig]

A list of steps to run.

rjfJobFlowRole :: Lens' RunJobFlow (Maybe Text)

Also called instance profile and EC2 role. An IAM role for an EMR cluster. The EC2 instances of the cluster assume this role. The default role is EMR_EC2_DefaultRole . In order to use the default role, you must have already created it using the CLI or console.

rjfBootstrapActions :: Lens' RunJobFlow [BootstrapActionConfig]

A list of bootstrap actions to run before Hadoop starts on the cluster nodes.

rjfReleaseLabel :: Lens' RunJobFlow (Maybe Text)

The release label for the Amazon EMR release. For Amazon EMR 3.x and 2.x AMIs, use AmiVersion instead.

rjfRepoUpgradeOnBoot :: Lens' RunJobFlow (Maybe RepoUpgradeOnBoot)

Applies only when CustomAmiID is used. Specifies which updates from the Amazon Linux AMI package repositories to apply automatically when the instance boots using the AMI. If omitted, the default is SECURITY , which indicates that only security updates are applied. If NONE is specified, no updates are applied, and all updates must be applied manually.

rjfLogURI :: Lens' RunJobFlow (Maybe Text)

The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

rjfNewSupportedProducts :: Lens' RunJobFlow [SupportedProductConfig]

A list of strings that indicates third-party software to use with the job flow that accepts a user argument list. EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action arguments. For more information, see "Launch a Job Flow on the MapR Distribution for Hadoop" in the Amazon EMR Developer Guide . Supported values are: * "mapr-m3" - launch the cluster using MapR M3 Edition. * "mapr-m5" - launch the cluster using MapR M5 Edition. * "mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition respectively. * "mapr-m7" - launch the cluster using MapR M7 Edition. * "hunk" - launch the cluster with the Hunk Big Data Analtics Platform. * "hue"- launch the cluster with Hue installed. * "spark" - launch the cluster with Apache Spark installed. * "ganglia" - launch the cluster with the Ganglia Monitoring System installed.

rjfVisibleToAllUsers :: Lens' RunJobFlow (Maybe Bool)

Whether the cluster is visible to all IAM users of the AWS account associated with the cluster. If this value is set to true , all IAM users of that AWS account can view and (if they have the proper policy permissions set) manage the cluster. If it is set to false , only the IAM user that created the cluster can view and manage it.

rjfSupportedProducts :: Lens' RunJobFlow [Text]

A list of strings that indicates third-party software to use. For more information, see Use Third Party Applications with Amazon EMR . Currently supported values are: * "mapr-m3" - launch the job flow using MapR M3 Edition. * "mapr-m5" - launch the job flow using MapR M5 Edition.

rjfApplications :: Lens' RunJobFlow [Application]

For Amazon EMR releases 4.0 and later. A list of applications for the cluster. Valid values are: Hadoop, Hive, Mahout, Pig, and "Spark." They are case insensitive.

rjfTags :: Lens' RunJobFlow [Tag]

A list of tags to associate with a cluster and propagate to Amazon EC2 instances.

rjfServiceRole :: Lens' RunJobFlow (Maybe Text)

The IAM role that will be assumed by the Amazon EMR service to access AWS resources on your behalf.

rjfName :: Lens' RunJobFlow Text

The name of the job flow.

rjfInstances :: Lens' RunJobFlow JobFlowInstancesConfig

A specification of the number and type of Amazon EC2 instances.

Destructuring the Response

runJobFlowResponse

Arguments

:: Int
rjfrsResponseStatus

-> RunJobFlowResponse

Creates a value of RunJobFlowResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rjfrsJobFlowId - An unique identifier for the job flow.

rjfrsResponseStatus - -- | The response status code.

data RunJobFlowResponse

The result of the RunJobFlow operation.

See: runJobFlowResponse smart constructor.

Response Lenses

rjfrsJobFlowId :: Lens' RunJobFlowResponse (Maybe Text)

An unique identifier for the job flow.

rjfrsResponseStatus :: Lens' RunJobFlowResponse Int

- | The response status code.