Network.AWS.EMR.RunJobFlow

Description

RunJobFlow creates and starts running a new job flow. The job flow will run the steps specified. Once the job flow completes, the cluster is stopped and the HDFS partition is lost. To prevent loss of data, configure the last step of the job flow to store results in Amazon S3. If the JobFlowInstancesConfig KeepJobFlowAliveWhenNoSteps parameter is set to TRUE, the job flow will transition to the WAITING state rather than shutting down once the steps have completed.

For additional protection, you can set the JobFlowInstancesConfig TerminationProtected parameter to TRUE to lock the job flow and prevent it from being terminated by API call, user intervention, or in the event of a job flow error.

A maximum of 256 steps are allowed in each job flow.

If your job flow is long-running (such as a Hive data warehouse) or complex, you may require more than 256 steps to process your data. You can bypass the 256-step limitation in various ways, including using the SSH shell to connect to the master node and submitting queries directly to the software running on the master node, such as Hive and Hadoop. For more information on how to do this, go to Add More than 256 Steps to a Job Flow in the Amazon ElasticMapReduce Developer's Guide.

For long running job flows, we recommend that you periodically store your results.

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_RunJobFlow.html

Request

data RunJobFlow

Request constructor

runJobFlow

Arguments

:: Text
rjfName

-> JobFlowInstancesConfig
rjfInstances

-> RunJobFlow

RunJobFlow constructor.

The fields accessible through corresponding lenses are:

rjfAdditionalInfo :: Maybe Text

rjfAmiVersion :: Maybe Text

rjfBootstrapActions :: [BootstrapActionConfig]

rjfInstances :: JobFlowInstancesConfig

rjfJobFlowRole :: Maybe Text

rjfLogUri :: Maybe Text

rjfName :: Text

rjfNewSupportedProducts :: [SupportedProductConfig]

rjfServiceRole :: Maybe Text

rjfSteps :: [StepConfig]

rjfSupportedProducts :: [Text]

rjfTags :: [Tag]

rjfVisibleToAllUsers :: Maybe Bool

Request lenses

rjfAdditionalInfo :: Lens' RunJobFlow (Maybe Text)

A JSON string for selecting additional features.

rjfAmiVersion :: Lens' RunJobFlow (Maybe Text)

The version of the Amazon Machine Image (AMI) to use when launching Amazon EC2 instances in the job flow. The following values are valid:

"latest" (uses the latest AMI) The version number of the AMI to use, for example, "2.0" If the AMI supports multiple versions of Hadoop (for example, AMI 1.0 supports both Hadoop 0.18 and 0.20) you can use the JobFlowInstancesConfig HadoopVersion parameter to modify the version of Hadoop from the defaults shown above.

For details about the AMI versions currently supported by Amazon Elastic MapReduce, go to AMI Versions Supported in Elastic MapReduce in the AmazonElastic MapReduce Developer's Guide.

rjfBootstrapActions :: Lens' RunJobFlow [BootstrapActionConfig]

A list of bootstrap actions that will be run before Hadoop is started on the cluster nodes.

rjfInstances :: Lens' RunJobFlow JobFlowInstancesConfig

A specification of the number and type of Amazon EC2 instances on which to run the job flow.

rjfJobFlowRole :: Lens' RunJobFlow (Maybe Text)

An IAM role for the job flow. The EC2 instances of the job flow assume this role. The default role is EMRJobflowDefault. In order to use the default role, you must have already created it using the CLI.

rjfLogUri :: Lens' RunJobFlow (Maybe Text)

The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

rjfName :: Lens' RunJobFlow Text

The name of the job flow.

rjfNewSupportedProducts :: Lens' RunJobFlow [SupportedProductConfig]

A list of strings that indicates third-party software to use with the job flow that accepts a user argument list. EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action arguments. For more information, see Launch a Job Flow on the MapR Distribution forHadoop. Currently supported values are:

"mapr-m3" - launch the job flow using MapR M3 Edition. "mapr-m5" - launch the job flow using MapR M5 Edition. "mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition respectively.

rjfServiceRole :: Lens' RunJobFlow (Maybe Text)

The IAM role that will be assumed by the Amazon EMR service to access AWS resources on your behalf.

rjfSteps :: Lens' RunJobFlow [StepConfig]

A list of steps to be executed by the job flow.

rjfSupportedProducts :: Lens' RunJobFlow [Text]

A list of strings that indicates third-party software to use with the job flow. For more information, go to Use Third Party Applications with Amazon EMR . Currently supported values are:

"mapr-m3" - launch the job flow using MapR M3 Edition. "mapr-m5" - launch the job flow using MapR M5 Edition.

rjfTags :: Lens' RunJobFlow [Tag]

A list of tags to associate with a cluster and propagate to Amazon EC2 instances.

rjfVisibleToAllUsers :: Lens' RunJobFlow (Maybe Bool)

Whether the job flow is visible to all IAM users of the AWS account associated with the job flow. If this value is set to true, all IAM users of that AWS account can view and (if they have the proper policy permissions set) manage the job flow. If it is set to false, only the IAM user that created the job flow can view and manage it.

Response

data RunJobFlowResponse

Response constructor

runJobFlowResponse :: RunJobFlowResponse

RunJobFlowResponse constructor.

The fields accessible through corresponding lenses are:

rjfrJobFlowId :: Maybe Text

Response lenses

rjfrJobFlowId :: Lens' RunJobFlowResponse (Maybe Text)

An unique identifier for the job flow.