Network.AWS.EMR.SetVisibleToAllUsers

Description

Sets whether all AWS Identity and Access Management (IAM) users under your account can access the specified job flows. This action works on running job flows. You can also set the visibility of a job flow when you launch it using the VisibleToAllUsers parameter of RunJobFlow. The SetVisibleToAllUsers action can be called only by an IAM user who created the job flow or the AWS account that owns the job flow.

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_SetVisibleToAllUsers.html

Request

data SetVisibleToAllUsers

Request constructor

setVisibleToAllUsers

Arguments

:: Bool
svtauVisibleToAllUsers

-> SetVisibleToAllUsers

SetVisibleToAllUsers constructor.

The fields accessible through corresponding lenses are:

svtauJobFlowIds :: [Text]

svtauVisibleToAllUsers :: Bool

Request lenses

svtauJobFlowIds :: Lens' SetVisibleToAllUsers [Text]

Identifiers of the job flows to receive the new visibility setting.

svtauVisibleToAllUsers :: Lens' SetVisibleToAllUsers Bool

Whether the specified job flows are visible to all IAM users of the AWS account associated with the job flow. If this value is set to True, all IAM users of that AWS account can view and, if they have the proper IAM policy permissions set, manage the job flows. If it is set to False, only the IAM user that created a job flow can view and manage it.

Response

data SetVisibleToAllUsersResponse

Response constructor

setVisibleToAllUsersResponse :: SetVisibleToAllUsersResponse

SetVisibleToAllUsersResponse constructor.