Network.AWS.EMR.TerminateJobFlows

Description

TerminateJobFlows shuts a list of job flows down. When a job flow is shut down, any step not yet completed is canceled and the EC2 instances on which the job flow is running are stopped. Any log files not already saved are uploaded to Amazon S3 if a LogUri was specified when the job flow was created.

The maximum number of JobFlows allowed is 10. The call to TerminateJobFlows is asynchronous. Depending on the configuration of the job flow, it may take up to 5-20 minutes for the job flow to completely terminate and release allocated resources, such as Amazon EC2 instances.

http://docs.aws.amazon.com/ElasticMapReduce/latest/API/API_TerminateJobFlows.html

Request

data TerminateJobFlows

Request constructor

terminateJobFlows :: TerminateJobFlows

TerminateJobFlows constructor.

The fields accessible through corresponding lenses are:

tjfJobFlowIds :: [Text]

Request lenses

tjfJobFlowIds :: Lens' TerminateJobFlows [Text]

A list of job flows to be shutdown.

Response

data TerminateJobFlowsResponse

Response constructor

terminateJobFlowsResponse :: TerminateJobFlowsResponse

TerminateJobFlowsResponse constructor.