Network.AWS.EMR.Types

Service

data EMR

Version 2009-03-31 of the Amazon Elastic MapReduce service.

Error

data JSONError :: *

MarketType

data MarketType

Constructors

OnDemand
ON_DEMAND

Spot
SPOT

InstanceGroupConfig

data InstanceGroupConfig

instanceGroupConfig

Arguments

:: InstanceRoleType
igcInstanceRole

-> Text
igcInstanceType

-> Int
igcInstanceCount

-> InstanceGroupConfig

InstanceGroupConfig constructor.

The fields accessible through corresponding lenses are:

igcBidPrice :: Maybe Text

igcInstanceCount :: Int

igcInstanceRole :: InstanceRoleType

igcInstanceType :: Text

igcMarket :: Maybe MarketType

igcName :: Maybe Text

igcBidPrice :: Lens' InstanceGroupConfig (Maybe Text)

Bid price for each Amazon EC2 instance in the instance group when launching nodes as Spot Instances, expressed in USD.

igcInstanceCount :: Lens' InstanceGroupConfig Int

Target number of instances for the instance group.

igcInstanceRole :: Lens' InstanceGroupConfig InstanceRoleType

The role of the instance group in the cluster.

igcInstanceType :: Lens' InstanceGroupConfig Text

The Amazon EC2 instance type for all instances in the instance group.

igcMarket :: Lens' InstanceGroupConfig (Maybe MarketType)

Market type of the Amazon EC2 instances used to create a cluster node.

igcName :: Lens' InstanceGroupConfig (Maybe Text)

Friendly name given to the instance group.

InstanceStateChangeReason

data InstanceStateChangeReason

instanceStateChangeReason :: InstanceStateChangeReason

InstanceStateChangeReason constructor.

The fields accessible through corresponding lenses are:

iscrCode :: Maybe InstanceStateChangeReasonCode

iscrMessage :: Maybe Text

iscrCode :: Lens' InstanceStateChangeReason (Maybe InstanceStateChangeReasonCode)

The programmable code for the state change reason.

iscrMessage :: Lens' InstanceStateChangeReason (Maybe Text)

The status change reason description.

JobFlowDetail

data JobFlowDetail

jobFlowDetail

Arguments

:: Text
jfdJobFlowId

-> Text
jfdName

-> JobFlowExecutionStatusDetail
jfdExecutionStatusDetail

-> JobFlowInstancesDetail
jfdInstances

-> JobFlowDetail

JobFlowDetail constructor.

The fields accessible through corresponding lenses are:

jfdAmiVersion :: Maybe Text

jfdBootstrapActions :: [BootstrapActionDetail]

jfdExecutionStatusDetail :: JobFlowExecutionStatusDetail

jfdInstances :: JobFlowInstancesDetail

jfdJobFlowId :: Text

jfdJobFlowRole :: Maybe Text

jfdLogUri :: Maybe Text

jfdName :: Text

jfdServiceRole :: Maybe Text

jfdSteps :: [StepDetail]

jfdSupportedProducts :: [Text]

jfdVisibleToAllUsers :: Maybe Bool

jfdAmiVersion :: Lens' JobFlowDetail (Maybe Text)

The version of the AMI used to initialize Amazon EC2 instances in the job flow. For a list of AMI versions currently supported by Amazon ElasticMapReduce, go to AMI Versions Supported in Elastic MapReduce in the Amazon Elastic MapReduce Developer's Guide.

jfdBootstrapActions :: Lens' JobFlowDetail [BootstrapActionDetail]

A list of the bootstrap actions run by the job flow.

jfdExecutionStatusDetail :: Lens' JobFlowDetail JobFlowExecutionStatusDetail

Describes the execution status of the job flow.

jfdInstances :: Lens' JobFlowDetail JobFlowInstancesDetail

Describes the Amazon EC2 instances of the job flow.

jfdJobFlowId :: Lens' JobFlowDetail Text

The job flow identifier.

jfdJobFlowRole :: Lens' JobFlowDetail (Maybe Text)

The IAM role that was specified when the job flow was launched. The EC2 instances of the job flow assume this role.

jfdLogUri :: Lens' JobFlowDetail (Maybe Text)

The location in Amazon S3 where log files for the job are stored.

jfdName :: Lens' JobFlowDetail Text

The name of the job flow.

jfdServiceRole :: Lens' JobFlowDetail (Maybe Text)

The IAM role that will be assumed by the Amazon EMR service to access AWS resources on your behalf.

jfdSteps :: Lens' JobFlowDetail [StepDetail]

A list of steps run by the job flow.

jfdSupportedProducts :: Lens' JobFlowDetail [Text]

A list of strings set by third party software when the job flow is launched. If you are not using third party software to manage the job flow this value is empty.

jfdVisibleToAllUsers :: Lens' JobFlowDetail (Maybe Bool)

Specifies whether the job flow is visible to all IAM users of the AWS account associated with the job flow. If this value is set to true, all IAM users of that AWS account can view and (if they have the proper policy permissions set) manage the job flow. If it is set to false, only the IAM user that created the job flow can view and manage it. This value can be changed using the SetVisibleToAllUsers action.

KeyValue

data KeyValue

keyValue :: KeyValue

KeyValue constructor.

The fields accessible through corresponding lenses are:

kvKey :: Maybe Text

kvValue :: Maybe Text

kvKey :: Lens' KeyValue (Maybe Text)

The unique identifier of a key value pair.

kvValue :: Lens' KeyValue (Maybe Text)

The value part of the identified key.

SupportedProductConfig

data SupportedProductConfig

supportedProductConfig :: SupportedProductConfig

SupportedProductConfig constructor.

The fields accessible through corresponding lenses are:

spcArgs :: [Text]

spcName :: Maybe Text

spcArgs :: Lens' SupportedProductConfig [Text]

The list of user-supplied arguments.

spcName :: Lens' SupportedProductConfig (Maybe Text)

The name of the product configuration.

Command

data Command

command :: Command

Command constructor.

The fields accessible through corresponding lenses are:

cArgs :: [Text]

cName :: Maybe Text

cScriptPath :: Maybe Text

cArgs :: Lens' Command [Text]

Arguments for Amazon EMR to pass to the command for execution.

cName :: Lens' Command (Maybe Text)

The name of the command.

cScriptPath :: Lens' Command (Maybe Text)

The Amazon S3 location of the command script.

StepExecutionState

data StepExecutionState

Constructors

Cancelled
CANCELLED

Completed
COMPLETED

Continue
CONTINUE

Failed
FAILED

Interrupted
INTERRUPTED

Pending
PENDING

Running
RUNNING

ActionOnFailure

data ActionOnFailure

Constructors

AOFCancelAndWait
CANCEL_AND_WAIT

AOFContinue
CONTINUE

AOFTerminateCluster
TERMINATE_CLUSTER

AOFTerminateJobFlow
TERMINATE_JOB_FLOW

ClusterStateChangeReason

data ClusterStateChangeReason

clusterStateChangeReason :: ClusterStateChangeReason

ClusterStateChangeReason constructor.

The fields accessible through corresponding lenses are:

cscrCode :: Maybe ClusterStateChangeReasonCode

cscrMessage :: Maybe Text

cscrCode :: Lens' ClusterStateChangeReason (Maybe ClusterStateChangeReasonCode)

The programmatic code for the state change reason.

cscrMessage :: Lens' ClusterStateChangeReason (Maybe Text)

The descriptive message for the state change reason.

Tag

data Tag

tag :: Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Maybe Text

tagValue :: Maybe Text

tagKey :: Lens' Tag (Maybe Text)

A user-defined key, which is the minimum required information for a valid tag. For more information, see Tagging Amazon EMR Resources.

tagValue :: Lens' Tag (Maybe Text)

A user-defined value, which is optional in a tag. For more information, see Tagging Amazon EMR Resources.

Application

data Application

application :: Application

Application constructor.

The fields accessible through corresponding lenses are:

aAdditionalInfo :: HashMap Text Text

aArgs :: [Text]

aName :: Maybe Text

aVersion :: Maybe Text

aAdditionalInfo :: Lens' Application (HashMap Text Text)

This option is for advanced users only. This is meta information about third-party applications that third-party vendors use for testing purposes.

aArgs :: Lens' Application [Text]

Arguments for Amazon EMR to pass to the application.

aName :: Lens' Application (Maybe Text)

The name of the application.

aVersion :: Lens' Application (Maybe Text)

The version of the application.

JobFlowExecutionStatusDetail

data JobFlowExecutionStatusDetail

jobFlowExecutionStatusDetail

Arguments

:: JobFlowExecutionState
jfesdState

-> UTCTime
jfesdCreationDateTime

-> JobFlowExecutionStatusDetail

JobFlowExecutionStatusDetail constructor.

The fields accessible through corresponding lenses are:

jfesdCreationDateTime :: UTCTime

jfesdEndDateTime :: Maybe UTCTime

jfesdLastStateChangeReason :: Maybe Text

jfesdReadyDateTime :: Maybe UTCTime

jfesdStartDateTime :: Maybe UTCTime

jfesdState :: JobFlowExecutionState

jfesdCreationDateTime :: Lens' JobFlowExecutionStatusDetail UTCTime

The creation date and time of the job flow.

jfesdEndDateTime :: Lens' JobFlowExecutionStatusDetail (Maybe UTCTime)

The completion date and time of the job flow.

jfesdLastStateChangeReason :: Lens' JobFlowExecutionStatusDetail (Maybe Text)

Description of the job flow last changed state.

jfesdReadyDateTime :: Lens' JobFlowExecutionStatusDetail (Maybe UTCTime)

The date and time when the job flow was ready to start running bootstrap actions.

jfesdStartDateTime :: Lens' JobFlowExecutionStatusDetail (Maybe UTCTime)

The start date and time of the job flow.

jfesdState :: Lens' JobFlowExecutionStatusDetail JobFlowExecutionState

The state of the job flow.

InstanceGroupStatus

data InstanceGroupStatus

instanceGroupStatus :: InstanceGroupStatus

InstanceGroupStatus constructor.

The fields accessible through corresponding lenses are:

igsState :: Maybe InstanceGroupState

igsStateChangeReason :: Maybe InstanceGroupStateChangeReason

igsTimeline :: Maybe InstanceGroupTimeline

igsState :: Lens' InstanceGroupStatus (Maybe InstanceGroupState)

The current state of the instance group.

igsStateChangeReason :: Lens' InstanceGroupStatus (Maybe InstanceGroupStateChangeReason)

The status change reason details for the instance group.

igsTimeline :: Lens' InstanceGroupStatus (Maybe InstanceGroupTimeline)

The timeline of the instance group status over time.

Cluster

data Cluster

cluster

Arguments

:: Text
c1Id

-> Text
c1Name

-> ClusterStatus
c1Status

-> Cluster

Cluster constructor.

The fields accessible through corresponding lenses are:

c1Applications :: [Application]

c1AutoTerminate :: Maybe Bool

c1Ec2InstanceAttributes :: Maybe Ec2InstanceAttributes

c1Id :: Text

c1LogUri :: Maybe Text

c1MasterPublicDnsName :: Maybe Text

c1Name :: Text

c1NormalizedInstanceHours :: Maybe Int

c1RequestedAmiVersion :: Maybe Text

c1RunningAmiVersion :: Maybe Text

c1ServiceRole :: Maybe Text

c1Status :: ClusterStatus

c1Tags :: [Tag]

c1TerminationProtected :: Maybe Bool

c1VisibleToAllUsers :: Maybe Bool

c1Applications :: Lens' Cluster [Application]

The applications installed on this cluster.

c1AutoTerminate :: Lens' Cluster (Maybe Bool)

Specifies whether the cluster should terminate after completing all steps.

c1Ec2InstanceAttributes :: Lens' Cluster (Maybe Ec2InstanceAttributes)

c1Id :: Lens' Cluster Text

The unique identifier for the cluster.

c1LogUri :: Lens' Cluster (Maybe Text)

The path to the Amazon S3 location where logs for this cluster are stored.

c1MasterPublicDnsName :: Lens' Cluster (Maybe Text)

The public DNS name of the master Ec2 instance.

c1Name :: Lens' Cluster Text

The name of the cluster.

c1NormalizedInstanceHours :: Lens' Cluster (Maybe Int)

An approximation of the cost of the job flow, represented in m1.small/hours. This value is incremented one time for every hour an m1.small instance runs. Larger instances are weighted more, so an EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

c1RequestedAmiVersion :: Lens' Cluster (Maybe Text)

The AMI version requested for this cluster.

c1RunningAmiVersion :: Lens' Cluster (Maybe Text)

The AMI version running on this cluster. This differs from the requested version only if the requested version is a meta version, such as "latest".

c1ServiceRole :: Lens' Cluster (Maybe Text)

The IAM role that will be assumed by the Amazon EMR service to access AWS resources on your behalf.

c1Status :: Lens' Cluster ClusterStatus

The current status details about the cluster.

c1Tags :: Lens' Cluster [Tag]

A list of tags associated with a cluster.

c1TerminationProtected :: Lens' Cluster (Maybe Bool)

Indicates whether Amazon EMR will lock the cluster to prevent the EC2 instances from being terminated by an API call or user intervention, or in the event of a cluster error.

c1VisibleToAllUsers :: Lens' Cluster (Maybe Bool)

Indicates whether the job flow is visible to all IAM users of the AWS account associated with the job flow. If this value is set to true, all IAM users of that AWS account can view and manage the job flow if they have the proper policy permissions set. If this value is false, only the IAM user that created the cluster can view and manage it. This value can be changed using the SetVisibleToAllUsers action.

InstanceTimeline

data InstanceTimeline

instanceTimeline :: InstanceTimeline

InstanceTimeline constructor.

The fields accessible through corresponding lenses are:

itCreationDateTime :: Maybe UTCTime

itEndDateTime :: Maybe UTCTime

itReadyDateTime :: Maybe UTCTime

itCreationDateTime :: Lens' InstanceTimeline (Maybe UTCTime)

The creation date and time of the instance.

itEndDateTime :: Lens' InstanceTimeline (Maybe UTCTime)

The date and time when the instance was terminated.

itReadyDateTime :: Lens' InstanceTimeline (Maybe UTCTime)

The date and time when the instance was ready to perform tasks.

Ec2InstanceAttributes

data Ec2InstanceAttributes

ec2InstanceAttributes :: Ec2InstanceAttributes

Ec2InstanceAttributes constructor.

The fields accessible through corresponding lenses are:

eiaAdditionalMasterSecurityGroups :: [Text]

eiaAdditionalSlaveSecurityGroups :: [Text]

eiaEc2AvailabilityZone :: Maybe Text

eiaEc2KeyName :: Maybe Text

eiaEc2SubnetId :: Maybe Text

eiaEmrManagedMasterSecurityGroup :: Maybe Text

eiaEmrManagedSlaveSecurityGroup :: Maybe Text

eiaIamInstanceProfile :: Maybe Text

eiaAdditionalMasterSecurityGroups :: Lens' Ec2InstanceAttributes [Text]

A list of additional Amazon EC2 security group IDs for the master node.

eiaAdditionalSlaveSecurityGroups :: Lens' Ec2InstanceAttributes [Text]

A list of additional Amazon EC2 security group IDs for the slave nodes.

eiaEc2AvailabilityZone :: Lens' Ec2InstanceAttributes (Maybe Text)

The Availability Zone in which the cluster will run.

eiaEc2KeyName :: Lens' Ec2InstanceAttributes (Maybe Text)

The name of the Amazon EC2 key pair to use when connecting with SSH into the master node as a user named "hadoop".

eiaEc2SubnetId :: Lens' Ec2InstanceAttributes (Maybe Text)

To launch the job flow in Amazon VPC, set this parameter to the identifier of the Amazon VPC subnet where you want the job flow to launch. If you do not specify this value, the job flow is launched in the normal AWS cloud, outside of a VPC.

Amazon VPC currently does not support cluster compute quadruple extra large (cc1.4xlarge) instances. Thus, you cannot specify the cc1.4xlarge instance type for nodes of a job flow launched in a VPC.

eiaEmrManagedMasterSecurityGroup :: Lens' Ec2InstanceAttributes (Maybe Text)

The identifier of the Amazon EC2 security group (managed by Amazon Elastic MapReduce) for the master node.

eiaEmrManagedSlaveSecurityGroup :: Lens' Ec2InstanceAttributes (Maybe Text)

The identifier of the Amazon EC2 security group (managed by Amazon Elastic MapReduce) for the slave nodes.

eiaIamInstanceProfile :: Lens' Ec2InstanceAttributes (Maybe Text)

The IAM role that was specified when the job flow was launched. The EC2 instances of the job flow assume this role.

StepStateChangeReasonCode

data StepStateChangeReasonCode

Constructors

None
NONE

ClusterState

data ClusterState

Constructors

CSBootstrapping
BOOTSTRAPPING

CSRunning
RUNNING

CSStarting
STARTING

CSTerminated
TERMINATED

CSTerminatedWithErrors
TERMINATED_WITH_ERRORS

CSTerminating
TERMINATING

CSWaiting
WAITING

HadoopStepConfig

data HadoopStepConfig

hadoopStepConfig :: HadoopStepConfig

HadoopStepConfig constructor.

The fields accessible through corresponding lenses are:

hscArgs :: [Text]

hscJar :: Maybe Text

hscMainClass :: Maybe Text

hscProperties :: HashMap Text Text

hscArgs :: Lens' HadoopStepConfig [Text]

The list of command line arguments to pass to the JAR file's main function for execution.

hscJar :: Lens' HadoopStepConfig (Maybe Text)

The path to the JAR file that runs during the step.

hscMainClass :: Lens' HadoopStepConfig (Maybe Text)

The name of the main class in the specified Java file. If not specified, the JAR file should specify a main class in its manifest file.

hscProperties :: Lens' HadoopStepConfig (HashMap Text Text)

The list of Java properties that are set when the step runs. You can use these properties to pass key value pairs to your main function.

JobFlowExecutionState

data JobFlowExecutionState

Constructors

JFESBootstrapping
BOOTSTRAPPING

JFESCompleted
COMPLETED

JFESFailed
FAILED

JFESRunning
RUNNING

JFESShuttingDown
SHUTTING_DOWN

JFESStarting
STARTING

JFESTerminated
TERMINATED

JFESWaiting
WAITING

StepDetail

data StepDetail

stepDetail

Arguments

:: StepConfig
sdStepConfig

-> StepExecutionStatusDetail
sdExecutionStatusDetail

-> StepDetail

StepDetail constructor.

The fields accessible through corresponding lenses are:

sdExecutionStatusDetail :: StepExecutionStatusDetail

sdStepConfig :: StepConfig

sdExecutionStatusDetail :: Lens' StepDetail StepExecutionStatusDetail

The description of the step status.

sdStepConfig :: Lens' StepDetail StepConfig

The step configuration.

InstanceGroupStateChangeReason

data InstanceGroupStateChangeReason

instanceGroupStateChangeReason :: InstanceGroupStateChangeReason

InstanceGroupStateChangeReason constructor.

The fields accessible through corresponding lenses are:

igscrCode :: Maybe InstanceGroupStateChangeReasonCode

igscrMessage :: Maybe Text

igscrCode :: Lens' InstanceGroupStateChangeReason (Maybe InstanceGroupStateChangeReasonCode)

The programmable code for the state change reason.

igscrMessage :: Lens' InstanceGroupStateChangeReason (Maybe Text)

The status change reason description.

InstanceGroupType

data InstanceGroupType

Constructors

Core
CORE

Master
MASTER

Task
TASK

InstanceGroupStateChangeReasonCode

data InstanceGroupStateChangeReasonCode

Constructors

ClusterTerminated
CLUSTER_TERMINATED

InstanceFailure
INSTANCE_FAILURE

InternalError
INTERNAL_ERROR

ValidationError
VALIDATION_ERROR

StepStatus

data StepStatus

stepStatus :: StepStatus

StepStatus constructor.

The fields accessible through corresponding lenses are:

ssState :: Maybe StepState

ssStateChangeReason :: Maybe StepStateChangeReason

ssTimeline :: Maybe StepTimeline

ssState :: Lens' StepStatus (Maybe StepState)

The execution state of the cluster step.

ssStateChangeReason :: Lens' StepStatus (Maybe StepStateChangeReason)

The reason for the step execution status change.

ssTimeline :: Lens' StepStatus (Maybe StepTimeline)

The timeline of the cluster step status over time.

StepSummary

data StepSummary

stepSummary :: StepSummary

StepSummary constructor.

The fields accessible through corresponding lenses are:

ssActionOnFailure :: Maybe ActionOnFailure

ssConfig :: Maybe HadoopStepConfig

ssId :: Maybe Text

ssName :: Maybe Text

ssStatus :: Maybe StepStatus

ssActionOnFailure :: Lens' StepSummary (Maybe ActionOnFailure)

This specifies what action to take when the cluster step fails. Possible values are TERMINATE_CLUSTER, CANCEL_AND_WAIT, and CONTINUE.

ssConfig :: Lens' StepSummary (Maybe HadoopStepConfig)

The Hadoop job configuration of the cluster step.

ssId :: Lens' StepSummary (Maybe Text)

The identifier of the cluster step.

ssName :: Lens' StepSummary (Maybe Text)

The name of the cluster step.

ssStatus :: Lens' StepSummary (Maybe StepStatus)

The current execution status details of the cluster step.

InstanceGroupState

data InstanceGroupState

Constructors

IGSArrested
ARRESTED

IGSBootstrapping
BOOTSTRAPPING

IGSEnded
ENDED

IGSProvisioning
PROVISIONING

IGSResizing
RESIZING

IGSRunning
RUNNING

IGSShuttingDown
SHUTTING_DOWN

IGSSuspended
SUSPENDED

IGSTerminated
TERMINATED

IGSTerminating
TERMINATING

StepTimeline

data StepTimeline

stepTimeline :: StepTimeline

StepTimeline constructor.

The fields accessible through corresponding lenses are:

stCreationDateTime :: Maybe UTCTime

stEndDateTime :: Maybe UTCTime

stStartDateTime :: Maybe UTCTime

stCreationDateTime :: Lens' StepTimeline (Maybe UTCTime)

The date and time when the cluster step was created.

stEndDateTime :: Lens' StepTimeline (Maybe UTCTime)

The date and time when the cluster step execution completed or failed.

stStartDateTime :: Lens' StepTimeline (Maybe UTCTime)

The date and time when the cluster step execution started.

PlacementType

data PlacementType

placementType

Arguments

:: Text
ptAvailabilityZone

-> PlacementType

PlacementType constructor.

The fields accessible through corresponding lenses are:

ptAvailabilityZone :: Text

ptAvailabilityZone :: Lens' PlacementType Text

The Amazon EC2 Availability Zone for the job flow.

HadoopJarStepConfig

data HadoopJarStepConfig

hadoopJarStepConfig

Arguments

:: Text
hjscJar

-> HadoopJarStepConfig

HadoopJarStepConfig constructor.

The fields accessible through corresponding lenses are:

hjscArgs :: [Text]

hjscJar :: Text

hjscMainClass :: Maybe Text

hjscProperties :: [KeyValue]

hjscArgs :: Lens' HadoopJarStepConfig [Text]

A list of command line arguments passed to the JAR file's main function when executed.

hjscJar :: Lens' HadoopJarStepConfig Text

A path to a JAR file run during the step.

hjscMainClass :: Lens' HadoopJarStepConfig (Maybe Text)

The name of the main class in the specified Java file. If not specified, the JAR file should specify a Main-Class in its manifest file.

hjscProperties :: Lens' HadoopJarStepConfig [KeyValue]

A list of Java properties that are set when the step runs. You can use these properties to pass key value pairs to your main function.

InstanceGroupModifyConfig

data InstanceGroupModifyConfig

instanceGroupModifyConfig

Arguments

:: Text
igmcInstanceGroupId

-> InstanceGroupModifyConfig

InstanceGroupModifyConfig constructor.

The fields accessible through corresponding lenses are:

igmcEC2InstanceIdsToTerminate :: [Text]

igmcInstanceCount :: Maybe Int

igmcInstanceGroupId :: Text

igmcEC2InstanceIdsToTerminate :: Lens' InstanceGroupModifyConfig [Text]

The EC2 InstanceIds to terminate. For advanced users only. Once you terminate the instances, the instance group will not return to its original requested size.

igmcInstanceCount :: Lens' InstanceGroupModifyConfig (Maybe Int)

Target size for the instance group.

igmcInstanceGroupId :: Lens' InstanceGroupModifyConfig Text

Unique ID of the instance group to expand or shrink.

InstanceGroupDetail

data InstanceGroupDetail

instanceGroupDetail

Arguments

:: MarketType
igdMarket

-> InstanceRoleType
igdInstanceRole

-> Text
igdInstanceType

-> Int
igdInstanceRequestCount

-> Int
igdInstanceRunningCount

-> InstanceGroupState
igdState

-> UTCTime
igdCreationDateTime

-> InstanceGroupDetail

InstanceGroupDetail constructor.

The fields accessible through corresponding lenses are:

igdBidPrice :: Maybe Text

igdCreationDateTime :: UTCTime

igdEndDateTime :: Maybe UTCTime

igdInstanceGroupId :: Maybe Text

igdInstanceRequestCount :: Int

igdInstanceRole :: InstanceRoleType

igdInstanceRunningCount :: Int

igdInstanceType :: Text

igdLastStateChangeReason :: Maybe Text

igdMarket :: MarketType

igdName :: Maybe Text

igdReadyDateTime :: Maybe UTCTime

igdStartDateTime :: Maybe UTCTime

igdState :: InstanceGroupState

igdBidPrice :: Lens' InstanceGroupDetail (Maybe Text)

Bid price for EC2 Instances when launching nodes as Spot Instances, expressed in USD.

igdCreationDateTime :: Lens' InstanceGroupDetail UTCTime

The date/time the instance group was created.

igdEndDateTime :: Lens' InstanceGroupDetail (Maybe UTCTime)

The date/time the instance group was terminated.

igdInstanceGroupId :: Lens' InstanceGroupDetail (Maybe Text)

Unique identifier for the instance group.

igdInstanceRequestCount :: Lens' InstanceGroupDetail Int

Target number of instances to run in the instance group.

igdInstanceRole :: Lens' InstanceGroupDetail InstanceRoleType

Instance group role in the cluster

igdInstanceRunningCount :: Lens' InstanceGroupDetail Int

Actual count of running instances.

igdInstanceType :: Lens' InstanceGroupDetail Text

Amazon EC2 Instance type.

igdLastStateChangeReason :: Lens' InstanceGroupDetail (Maybe Text)

Details regarding the state of the instance group.

igdMarket :: Lens' InstanceGroupDetail MarketType

Market type of the Amazon EC2 instances used to create a cluster node.

igdName :: Lens' InstanceGroupDetail (Maybe Text)

Friendly name for the instance group.

igdReadyDateTime :: Lens' InstanceGroupDetail (Maybe UTCTime)

The date/time the instance group was available to the cluster.

igdStartDateTime :: Lens' InstanceGroupDetail (Maybe UTCTime)

The date/time the instance group was started.

igdState :: Lens' InstanceGroupDetail InstanceGroupState

State of instance group. The following values are deprecated: STARTING, TERMINATED, and FAILED.

StepStateChangeReason

data StepStateChangeReason

stepStateChangeReason :: StepStateChangeReason

StepStateChangeReason constructor.

The fields accessible through corresponding lenses are:

sscrCode :: Maybe StepStateChangeReasonCode

sscrMessage :: Maybe Text

sscrCode :: Lens' StepStateChangeReason (Maybe StepStateChangeReasonCode)

The programmable code for the state change reason.

sscrMessage :: Lens' StepStateChangeReason (Maybe Text)

The descriptive message for the state change reason.

ClusterStateChangeReasonCode

data ClusterStateChangeReasonCode

Constructors

CSCRCAllStepsCompleted
ALL_STEPS_COMPLETED

CSCRCBootstrapFailure
BOOTSTRAP_FAILURE

CSCRCInstanceFailure
INSTANCE_FAILURE

CSCRCInternalError
INTERNAL_ERROR

CSCRCStepFailure
STEP_FAILURE

CSCRCUserRequest
USER_REQUEST

CSCRCValidationError
VALIDATION_ERROR

Step

data Step

step :: Step

Step constructor.

The fields accessible through corresponding lenses are:

sActionOnFailure :: Maybe ActionOnFailure

sConfig :: Maybe HadoopStepConfig

sId :: Maybe Text

sName :: Maybe Text

sStatus :: Maybe StepStatus

sActionOnFailure :: Lens' Step (Maybe ActionOnFailure)

This specifies what action to take when the cluster step fails. Possible values are TERMINATE_CLUSTER, CANCEL_AND_WAIT, and CONTINUE.

sConfig :: Lens' Step (Maybe HadoopStepConfig)

The Hadoop job configuration of the cluster step.

sId :: Lens' Step (Maybe Text)

The identifier of the cluster step.

sName :: Lens' Step (Maybe Text)

The name of the cluster step.

sStatus :: Lens' Step (Maybe StepStatus)

The current execution status details of the cluster step.

StepState

data StepState

Constructors

SSCancelled
CANCELLED

SSCompleted
COMPLETED

SSFailed
FAILED

SSInterrupted
INTERRUPTED

SSPending
PENDING

SSRunning
RUNNING

InstanceGroupTimeline

data InstanceGroupTimeline

instanceGroupTimeline :: InstanceGroupTimeline

InstanceGroupTimeline constructor.

The fields accessible through corresponding lenses are:

igtCreationDateTime :: Maybe UTCTime

igtEndDateTime :: Maybe UTCTime

igtReadyDateTime :: Maybe UTCTime

igtCreationDateTime :: Lens' InstanceGroupTimeline (Maybe UTCTime)

The creation date and time of the instance group.

igtEndDateTime :: Lens' InstanceGroupTimeline (Maybe UTCTime)

The date and time when the instance group terminated.

igtReadyDateTime :: Lens' InstanceGroupTimeline (Maybe UTCTime)

The date and time when the instance group became ready to perform tasks.

BootstrapActionDetail

data BootstrapActionDetail

bootstrapActionDetail :: BootstrapActionDetail

BootstrapActionDetail constructor.

The fields accessible through corresponding lenses are:

badBootstrapActionConfig :: Maybe BootstrapActionConfig

badBootstrapActionConfig :: Lens' BootstrapActionDetail (Maybe BootstrapActionConfig)

A description of the bootstrap action.

StepExecutionStatusDetail

data StepExecutionStatusDetail

stepExecutionStatusDetail

Arguments

:: StepExecutionState
sesdState

-> UTCTime
sesdCreationDateTime

-> StepExecutionStatusDetail

StepExecutionStatusDetail constructor.

The fields accessible through corresponding lenses are:

sesdCreationDateTime :: UTCTime

sesdEndDateTime :: Maybe UTCTime

sesdLastStateChangeReason :: Maybe Text

sesdStartDateTime :: Maybe UTCTime

sesdState :: StepExecutionState

sesdCreationDateTime :: Lens' StepExecutionStatusDetail UTCTime

The creation date and time of the step.

sesdEndDateTime :: Lens' StepExecutionStatusDetail (Maybe UTCTime)

The completion date and time of the step.

sesdLastStateChangeReason :: Lens' StepExecutionStatusDetail (Maybe Text)

A description of the step's current state.

sesdStartDateTime :: Lens' StepExecutionStatusDetail (Maybe UTCTime)

The start date and time of the step.

sesdState :: Lens' StepExecutionStatusDetail StepExecutionState

The state of the job flow step.

InstanceStatus

data InstanceStatus

instanceStatus :: InstanceStatus

InstanceStatus constructor.

The fields accessible through corresponding lenses are:

isState :: Maybe InstanceState

isStateChangeReason :: Maybe InstanceStateChangeReason

isTimeline :: Maybe InstanceTimeline

isState :: Lens' InstanceStatus (Maybe InstanceState)

The current state of the instance.

isStateChangeReason :: Lens' InstanceStatus (Maybe InstanceStateChangeReason)

The details of the status change reason for the instance.

isTimeline :: Lens' InstanceStatus (Maybe InstanceTimeline)

The timeline of the instance status over time.

InstanceRoleType

data InstanceRoleType

Constructors

IRTCore
CORE

IRTMaster
MASTER

IRTTask
TASK

JobFlowInstancesConfig

data JobFlowInstancesConfig

jobFlowInstancesConfig :: JobFlowInstancesConfig

JobFlowInstancesConfig constructor.

The fields accessible through corresponding lenses are:

jficAdditionalMasterSecurityGroups :: [Text]

jficAdditionalSlaveSecurityGroups :: [Text]

jficEc2KeyName :: Maybe Text

jficEc2SubnetId :: Maybe Text

jficEmrManagedMasterSecurityGroup :: Maybe Text

jficEmrManagedSlaveSecurityGroup :: Maybe Text

jficHadoopVersion :: Maybe Text

jficInstanceCount :: Maybe Int

jficInstanceGroups :: [InstanceGroupConfig]

jficKeepJobFlowAliveWhenNoSteps :: Maybe Bool

jficMasterInstanceType :: Maybe Text

jficPlacement :: Maybe PlacementType

jficSlaveInstanceType :: Maybe Text

jficTerminationProtected :: Maybe Bool

jficAdditionalMasterSecurityGroups :: Lens' JobFlowInstancesConfig [Text]

A list of additional Amazon EC2 security group IDs for the master node.

jficAdditionalSlaveSecurityGroups :: Lens' JobFlowInstancesConfig [Text]

A list of additional Amazon EC2 security group IDs for the slave nodes.

jficEc2KeyName :: Lens' JobFlowInstancesConfig (Maybe Text)

The name of the Amazon EC2 key pair that can be used to ssh to the master node as the user called "hadoop."

jficEc2SubnetId :: Lens' JobFlowInstancesConfig (Maybe Text)

To launch the job flow in Amazon Virtual Private Cloud (Amazon VPC), set this parameter to the identifier of the Amazon VPC subnet where you want the job flow to launch. If you do not specify this value, the job flow is launched in the normal Amazon Web Services cloud, outside of an Amazon VPC.

Amazon VPC currently does not support cluster compute quadruple extra large (cc1.4xlarge) instances. Thus you cannot specify the cc1.4xlarge instance type for nodes of a job flow launched in a Amazon VPC.

jficEmrManagedMasterSecurityGroup :: Lens' JobFlowInstancesConfig (Maybe Text)

The identifier of the Amazon EC2 security group (managed by Amazon ElasticMapReduce) for the master node.

jficEmrManagedSlaveSecurityGroup :: Lens' JobFlowInstancesConfig (Maybe Text)

The identifier of the Amazon EC2 security group (managed by Amazon ElasticMapReduce) for the slave nodes.

jficHadoopVersion :: Lens' JobFlowInstancesConfig (Maybe Text)

The Hadoop version for the job flow. Valid inputs are "0.18", "0.20", "0.20.205", "1.0.3", "2.2.0", or "2.4.0". If you do not set this value, the default of 0.18 is used, unless the AmiVersion parameter is set in the RunJobFlow call, in which case the default version of Hadoop for that AMI version is used.

jficInstanceCount :: Lens' JobFlowInstancesConfig (Maybe Int)

The number of Amazon EC2 instances used to execute the job flow.

jficInstanceGroups :: Lens' JobFlowInstancesConfig [InstanceGroupConfig]

Configuration for the job flow's instance groups.

jficKeepJobFlowAliveWhenNoSteps :: Lens' JobFlowInstancesConfig (Maybe Bool)

Specifies whether the job flow should terminate after completing all steps.

jficMasterInstanceType :: Lens' JobFlowInstancesConfig (Maybe Text)

The EC2 instance type of the master node.

jficPlacement :: Lens' JobFlowInstancesConfig (Maybe PlacementType)

The Availability Zone the job flow will run in.

jficSlaveInstanceType :: Lens' JobFlowInstancesConfig (Maybe Text)

The EC2 instance type of the slave nodes.

jficTerminationProtected :: Lens' JobFlowInstancesConfig (Maybe Bool)

Specifies whether to lock the job flow to prevent the Amazon EC2 instances from being terminated by API call, user intervention, or in the event of a job flow error.

StepConfig

data StepConfig

stepConfig

Arguments

:: Text
scName

-> HadoopJarStepConfig
scHadoopJarStep

-> StepConfig

StepConfig constructor.

The fields accessible through corresponding lenses are:

scActionOnFailure :: Maybe ActionOnFailure

scHadoopJarStep :: HadoopJarStepConfig

scName :: Text

scActionOnFailure :: Lens' StepConfig (Maybe ActionOnFailure)

The action to take if the job flow step fails.

scHadoopJarStep :: Lens' StepConfig HadoopJarStepConfig

The JAR file used for the job flow step.

scName :: Lens' StepConfig Text

The name of the job flow step.

InstanceGroup

data InstanceGroup

instanceGroup :: InstanceGroup

InstanceGroup constructor.

The fields accessible through corresponding lenses are:

igBidPrice :: Maybe Text

igId :: Maybe Text

igInstanceGroupType :: Maybe InstanceGroupType

igInstanceType :: Maybe Text

igMarket :: Maybe MarketType

igName :: Maybe Text

igRequestedInstanceCount :: Maybe Int

igRunningInstanceCount :: Maybe Int

igStatus :: Maybe InstanceGroupStatus

igBidPrice :: Lens' InstanceGroup (Maybe Text)

The bid price for each EC2 instance in the instance group when launching nodes as Spot Instances, expressed in USD.

igId :: Lens' InstanceGroup (Maybe Text)

The identifier of the instance group.

igInstanceGroupType :: Lens' InstanceGroup (Maybe InstanceGroupType)

The type of the instance group. Valid values are MASTER, CORE or TASK.

igInstanceType :: Lens' InstanceGroup (Maybe Text)

The EC2 instance type for all instances in the instance group.

igMarket :: Lens' InstanceGroup (Maybe MarketType)

The marketplace to provision instances for this group. Valid values are ON_DEMAND or SPOT.

igName :: Lens' InstanceGroup (Maybe Text)

The name of the instance group.

igRequestedInstanceCount :: Lens' InstanceGroup (Maybe Int)

The target number of instances for the instance group.

igRunningInstanceCount :: Lens' InstanceGroup (Maybe Int)

The number of instances currently running in this instance group.

igStatus :: Lens' InstanceGroup (Maybe InstanceGroupStatus)

The current status of the instance group.

BootstrapActionConfig

data BootstrapActionConfig

bootstrapActionConfig

Arguments

:: Text
bacName

-> ScriptBootstrapActionConfig
bacScriptBootstrapAction

-> BootstrapActionConfig

BootstrapActionConfig constructor.

The fields accessible through corresponding lenses are:

bacName :: Text

bacScriptBootstrapAction :: ScriptBootstrapActionConfig

bacName :: Lens' BootstrapActionConfig Text

The name of the bootstrap action.

bacScriptBootstrapAction :: Lens' BootstrapActionConfig ScriptBootstrapActionConfig

The script run by the bootstrap action.

ClusterSummary

data ClusterSummary

clusterSummary :: ClusterSummary

ClusterSummary constructor.

The fields accessible through corresponding lenses are:

csId :: Maybe Text

csName :: Maybe Text

csNormalizedInstanceHours :: Maybe Int

csStatus :: Maybe ClusterStatus

csId :: Lens' ClusterSummary (Maybe Text)

The unique identifier for the cluster.

csName :: Lens' ClusterSummary (Maybe Text)

The name of the cluster.

csNormalizedInstanceHours :: Lens' ClusterSummary (Maybe Int)

An approximation of the cost of the job flow, represented in m1.small/hours. This value is incremented one time for every hour an m1.small instance runs. Larger instances are weighted more, so an EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

csStatus :: Lens' ClusterSummary (Maybe ClusterStatus)

The details about the current status of the cluster.

JobFlowInstancesDetail

data JobFlowInstancesDetail

jobFlowInstancesDetail

Arguments

:: Text
jfidMasterInstanceType

-> Text
jfidSlaveInstanceType

-> Int
jfidInstanceCount

-> JobFlowInstancesDetail

JobFlowInstancesDetail constructor.

The fields accessible through corresponding lenses are:

jfidEc2KeyName :: Maybe Text

jfidEc2SubnetId :: Maybe Text

jfidHadoopVersion :: Maybe Text

jfidInstanceCount :: Int

jfidInstanceGroups :: [InstanceGroupDetail]

jfidKeepJobFlowAliveWhenNoSteps :: Maybe Bool

jfidMasterInstanceId :: Maybe Text

jfidMasterInstanceType :: Text

jfidMasterPublicDnsName :: Maybe Text

jfidNormalizedInstanceHours :: Maybe Int

jfidPlacement :: Maybe PlacementType

jfidSlaveInstanceType :: Text

jfidTerminationProtected :: Maybe Bool

jfidEc2KeyName :: Lens' JobFlowInstancesDetail (Maybe Text)

The name of an Amazon EC2 key pair that can be used to ssh to the master node of job flow.

jfidEc2SubnetId :: Lens' JobFlowInstancesDetail (Maybe Text)

For job flows launched within Amazon Virtual Private Cloud, this value specifies the identifier of the subnet where the job flow was launched.

jfidHadoopVersion :: Lens' JobFlowInstancesDetail (Maybe Text)

The Hadoop version for the job flow.

jfidInstanceCount :: Lens' JobFlowInstancesDetail Int

The number of Amazon EC2 instances in the cluster. If the value is 1, the same instance serves as both the master and slave node. If the value is greater than 1, one instance is the master node and all others are slave nodes.

jfidInstanceGroups :: Lens' JobFlowInstancesDetail [InstanceGroupDetail]

Details about the job flow's instance groups.

jfidKeepJobFlowAliveWhenNoSteps :: Lens' JobFlowInstancesDetail (Maybe Bool)

Specifies whether the job flow should terminate after completing all steps.

jfidMasterInstanceId :: Lens' JobFlowInstancesDetail (Maybe Text)

The Amazon EC2 instance identifier of the master node.

jfidMasterInstanceType :: Lens' JobFlowInstancesDetail Text

The Amazon EC2 master node instance type.

jfidMasterPublicDnsName :: Lens' JobFlowInstancesDetail (Maybe Text)

The DNS name of the master node.

jfidNormalizedInstanceHours :: Lens' JobFlowInstancesDetail (Maybe Int)

An approximation of the cost of the job flow, represented in m1.small/hours. This value is incremented once for every hour an m1.small runs. Larger instances are weighted more, so an Amazon EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

jfidPlacement :: Lens' JobFlowInstancesDetail (Maybe PlacementType)

The Amazon EC2 Availability Zone for the job flow.

jfidSlaveInstanceType :: Lens' JobFlowInstancesDetail Text

The Amazon EC2 slave node instance type.

jfidTerminationProtected :: Lens' JobFlowInstancesDetail (Maybe Bool)

Specifies whether the Amazon EC2 instances in the cluster are protected from termination by API calls, user intervention, or in the event of a job flow error.

ClusterStatus

data ClusterStatus

clusterStatus :: ClusterStatus

ClusterStatus constructor.

The fields accessible through corresponding lenses are:

csState :: Maybe ClusterState

csStateChangeReason :: Maybe ClusterStateChangeReason

csTimeline :: Maybe ClusterTimeline

csState :: Lens' ClusterStatus (Maybe ClusterState)

The current state of the cluster.

csStateChangeReason :: Lens' ClusterStatus (Maybe ClusterStateChangeReason)

The reason for the cluster status change.

csTimeline :: Lens' ClusterStatus (Maybe ClusterTimeline)

A timeline that represents the status of a cluster over the lifetime of the cluster.

InstanceState

data InstanceState

Constructors

ISAwaitingFulfillment
AWAITING_FULFILLMENT

ISBootstrapping
BOOTSTRAPPING

ISProvisioning
PROVISIONING

ISRunning
RUNNING

ISTerminated
TERMINATED

ClusterTimeline

data ClusterTimeline

clusterTimeline :: ClusterTimeline

ClusterTimeline constructor.

The fields accessible through corresponding lenses are:

ctCreationDateTime :: Maybe UTCTime

ctEndDateTime :: Maybe UTCTime

ctReadyDateTime :: Maybe UTCTime

ctCreationDateTime :: Lens' ClusterTimeline (Maybe UTCTime)

The creation date and time of the cluster.

ctEndDateTime :: Lens' ClusterTimeline (Maybe UTCTime)

The date and time when the cluster was terminated.

ctReadyDateTime :: Lens' ClusterTimeline (Maybe UTCTime)

The date and time when the cluster was ready to execute steps.

InstanceStateChangeReasonCode

data InstanceStateChangeReasonCode

Constructors

ISCRCBootstrapFailure
BOOTSTRAP_FAILURE

ISCRCClusterTerminated
CLUSTER_TERMINATED

ISCRCInstanceFailure
INSTANCE_FAILURE

ISCRCInternalError
INTERNAL_ERROR

ISCRCValidationError
VALIDATION_ERROR

Instance

data Instance

instance' :: Instance

Instance constructor.

The fields accessible through corresponding lenses are:

iEc2InstanceId :: Maybe Text

iId :: Maybe Text

iPrivateDnsName :: Maybe Text

iPrivateIpAddress :: Maybe Text

iPublicDnsName :: Maybe Text

iPublicIpAddress :: Maybe Text

iStatus :: Maybe InstanceStatus

iEc2InstanceId :: Lens' Instance (Maybe Text)

The unique identifier of the instance in Amazon EC2.

iId :: Lens' Instance (Maybe Text)

The unique identifier for the instance in Amazon EMR.

iPrivateDnsName :: Lens' Instance (Maybe Text)

The private DNS name of the instance.

iPrivateIpAddress :: Lens' Instance (Maybe Text)

The private IP address of the instance.

iPublicDnsName :: Lens' Instance (Maybe Text)

The public DNS name of the instance.

iPublicIpAddress :: Lens' Instance (Maybe Text)

The public IP address of the instance.

iStatus :: Lens' Instance (Maybe InstanceStatus)

The current status of the instance.

ScriptBootstrapActionConfig

data ScriptBootstrapActionConfig

scriptBootstrapActionConfig

Arguments

:: Text
sbacPath

-> ScriptBootstrapActionConfig

ScriptBootstrapActionConfig constructor.

The fields accessible through corresponding lenses are:

sbacArgs :: [Text]

sbacPath :: Text

sbacArgs :: Lens' ScriptBootstrapActionConfig [Text]

A list of command line arguments to pass to the bootstrap action script.

sbacPath :: Lens' ScriptBootstrapActionConfig Text

Location of the script to run during a bootstrap action. Can be either a location in Amazon S3 or on a local file system.