Network.AWS.EMR.V2009_03_31.AddInstanceGroups

Description

AddInstanceGroups adds an instance group to a running cluster. POST / HTTP1.1 Content-Type: applicationx-amz-json-1.1 X-Amz-Target: ElasticMapReduce.AddInstanceGroups Content-Length: 168 User-Agent: aws-sdk-ruby1.9.2 ruby1.9.3 i386-mingw32 Host: us-east-1.elasticmapreduce.amazonaws.com X-Amz-Date: 20130715T223346Z X-Amz-Content-Sha256: ac5a7193b1283898dd822a4b16ca36963879bb010d2dbe57198439973ab2a7d3 Authorization: AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20130715us-east-1elasticmapreduceaws4_request,

SignedHeaders=content-length;content-type;host;user-agent;x-amz-content-sha256;x-amz-date;x-amz-target, Signature=4c5e7eb762ea45f292a5cd1a1cc56ed60009e19a9dba3d6e5e4e67e96d43af11 Accept: */* { JobFlowId: "j-3U7TSX5GZFD8Y", InstanceGroups: [{ Name: "Task Instance Group", InstanceRole: TASK, InstanceCount: 2, InstanceType: "m1.small", Market: ON_DEMAND }] } HTTP/1.1 200 OK x-amzn-RequestId: 9da5a349-ed9e-11e2-90db-69a5154aeb8d Content-Type: application/x-amz-json-1.1 Content-Length: 71 Date: Mon, 15 Jul 2013 22:33:47 GMT { InstanceGroupIds: ["ig-294A6A2KWT4WB"], JobFlowId: "j-3U7TSX5GZFD8Y" }.

Request

data AddInstanceGroups

Input to an AddInstanceGroups call.

Request constructor

mkAddInstanceGroups

Arguments

:: [InstanceGroupConfig]
aigInstanceGroups

-> Text
aigJobFlowId

-> AddInstanceGroups

Smart constructor for the minimum required parameters to construct a valid AddInstanceGroups request.

Request lenses

aigInstanceGroups :: Lens' AddInstanceGroups [InstanceGroupConfig]

Instance Groups to add.

aigJobFlowId :: Lens' AddInstanceGroups Text

Job flow in which to add the instance groups.

Response

data AddInstanceGroupsResponse

Output from an AddInstanceGroups call.

Response constructor

mkAddInstanceGroupsResponse :: AddInstanceGroupsResponse

Smart constructor for the minimum required parameters to construct a valid AddInstanceGroupsResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

aigrJobFlowId :: Lens' AddInstanceGroupsResponse (Maybe Text)

The job flow ID in which the instance groups are added.

aigrInstanceGroupIds :: Lens' AddInstanceGroupsResponse [Text]

Instance group IDs of the newly created instance groups.