Network.AWS.EMR.V2009_03_31.AddTags

Description

Adds tags to an Amazon EMR resource. Tags make it easier to associate clusters in various ways, such as grouping clusters to track your Amazon EMR resource allocation costs. For more information, see Tagging Amazon EMR Resources. POST HTTP1.1 Content-Type: application/x-amz-json-1.1 X-Amz-Target: ElasticMapReduce.AddTags AUTHPARAMS { ResourceId: "j-3U7TSX5GZFD8Y", Tags: [{ Key: "stack", Value: Production }, { Key: "hbase" }] } HTTP/1.1 200 OK x-amzn-RequestId: 9da5a349-ed9e-11e2-90db-69a5154aeb8d Content-Type: application/x-amz-json-1.1 Content-Length: 71 Date: Mon, 15 Jul 2013 22:33:47 GMT { }.

Documentation

data AddTags

Constructors

AddTags

Fields

_atiResourceId :: Text

The Amazon EMR resource identifier to which tags will be added. This value must be a cluster identifier.

_atiTags :: [Tag]

A list of tags to associate with a cluster and propagate to Amazon EC2 instances. Tags are user-defined key/value pairs that consist of a required key string with a maximum of 128 characters, and an optional value string with a maximum of 256 characters.

atiTags :: Lens' AddTags [Tag]

atiResourceId :: Lens' AddTags Text

data AddTagsResponse

Constructors

AddTagsResponse